feat: Phase 4 - Initial code structure reorganization (WIP)
## Code Organization Improvements

### New Memory Management Module
- Create structured memory/ module with specialized submodules:
  - aligned.rs: SIMD-aligned memory allocation (moved from simd_aligned.rs)
  - optimization.rs: Memory optimization traits and implementations
  - pool.rs: Advanced memory pooling strategies
- Extract TensorMemoryInfo and MemoryOptimization trait from core.rs

### Zero-Copy Operations Abstraction
- Create operations/ module for organized tensor operations
- Extract zero-copy operations into trait-based system:
  - ZeroCopyOps trait: in-place operations, slicing, memory sharing
  - TensorIterOps trait: iterator operations for tensors
- Move implementation details out of core.rs for better separation

### Core Module Refactoring
- Remove duplicated code from core.rs (300+ lines reduced)
- Delegate memory optimization to MemoryOptimization trait
- Delegate zero-copy operations to ZeroCopyOps trait
- Cleaner imports using new module structure

### Module System Improvements
- Better organization of tensor functionality by domain
- Improved re-exports for user-friendly API
- Preparation for future modular expansions

## Technical Details

### Trait-Based Architecture
- MemoryOptimization<T>: memory_info(), can_optimize_memory(), optimize_memory()
- ZeroCopyOps<T>: inplace_add(), slice_view(), shares_memory_with(), detach()
- TensorIterOps<T>: iter(), iter_mut() with proper lifetime management

### API Preservation
- All existing functionality preserved through trait implementations
- No breaking changes to public API
- Backward compatibility maintained

## Status: Work in Progress
This represents initial structural improvements with more refinements needed.
The new architecture provides better separation of concerns and maintainability
while preserving all existing functionality.
