+++ src/tensor/core.rs
@@ -2,34 +2,9 @@
 //! コアテンソルデータ構造と基本操作
 
 use crate::error::{RusTorchError, RusTorchResult};
-use super::simd_aligned::{SimdAllocator, SIMD_ALIGNMENT};
-
-/// Memory information for tensors
-/// テンソルのメモリ情報
-#[derive(Debug, Clone)]
-pub struct TensorMemoryInfo {
-    /// Total number of elements in the tensor
-    /// テンソルの総要素数
-    pub total_elements: usize,
-    /// Size of each element in bytes
-    /// 各要素のバイトサイズ
-    pub element_size: usize,
-    /// Total memory used in bytes
-    /// 使用メモリの総バイト数
-    pub total_bytes: usize,
-    /// Whether the tensor data is contiguous in memory
-    /// テンソルデータがメモリ上で連続しているか
-    pub is_contiguous: bool,
-    /// Memory alignment in bytes
-    /// メモリアライメント（バイト）
-    pub alignment: usize,
-    /// Whether the tensor is stored on GPU
-    /// テンソルがGPU上に保存されているか
-    pub is_on_gpu: bool,
-    /// Device type string
-    /// デバイスタイプ文字列
-    pub device: String,
-}
+use super::memory::aligned::{SimdAllocator, SIMD_ALIGNMENT};
+use super::memory::optimization::{MemoryOptimization, TensorMemoryInfo};
+use super::operations::zero_copy::{ZeroCopyOps, TensorIterOps};
 use ndarray::{ArrayD, IxDyn};
 use num_traits::Float;
 use std::fmt;
@@ -185,113 +160,8 @@ impl<T: Float + 'static> Tensor<T> {
         false // Default implementation
     }
 
-    /// Get memory usage statistics for this tensor
-    /// このテンソルのメモリ使用量統計を取得
-    pub fn memory_info(&self) -> TensorMemoryInfo {
-        let element_size = std::mem::size_of::<T>();
-        let total_elements = self.data.len();
-        let total_bytes = total_elements * element_size;
-        
-        // Check for SIMD alignment
-        let ptr = self.data.as_ptr();
-        let alignment = if SimdAllocator::is_aligned(ptr) {
-            SIMD_ALIGNMENT
-        } else {
-            // Check for standard alignments
-            if (ptr as usize) % 16 == 0 {
-                16
-            } else if (ptr as usize) % 8 == 0 {
-                8
-            } else if (ptr as usize) % 4 == 0 {
-                4
-            } else {
-                1
-            }
-        };
-        
-        let is_on_gpu = self.is_on_gpu();
-        let device = self.device_type().to_string();
-        
-        TensorMemoryInfo {
-            total_elements,
-            element_size,
-            total_bytes,
-            is_contiguous: self.data.is_standard_layout(),
-            alignment,
-            is_on_gpu,
-            device,
-        }
-    }
-
-    /// Check if this tensor can be optimized for memory usage
-    /// このテンソルがメモリ使用量を最適化できるかチェック
-    pub fn can_optimize_memory(&self) -> bool {
-        let info = self.memory_info();
-        // Can optimize if tensor is large enough and not already SIMD-aligned
-        info.total_bytes > 1024 && info.alignment < SIMD_ALIGNMENT
-    }
-
-    /// Create a memory-optimized copy of this tensor
-    /// このテンソルのメモリ最適化コピーを作成
-    pub fn optimize_memory(&self) -> Self {
-        if !self.can_optimize_memory() {
-            return self.clone();
-        }
-        
-        // Try to create SIMD-aligned copy for f32 tensors
-        if std::any::TypeId::of::<T>() == std::any::TypeId::of::<f32>() {
-            let shape = self.shape();
-            let len = self.numel();
-            
-            // Allocate SIMD-aligned memory
-            if let Ok(ptr) = SimdAllocator::alloc_f32(len) {
-                unsafe {
-                    // Copy data to aligned memory
-                    let src = self.data.as_ptr();
-                    let dst = ptr.as_ptr();
-                    std::ptr::copy_nonoverlapping(src as *const f32, dst, len);
-                    
-                    // Create vector from aligned memory
-                    let aligned_data = Vec::from_raw_parts(dst, len, len);
-                    
-                    // Convert to T type (this is safe because we checked TypeId)
-                    let aligned_data_t: Vec<T> = std::mem::transmute(aligned_data);
-                    
-                    // Create tensor from aligned data
-                    match Self::try_from_vec(aligned_data_t, shape.to_vec()) {
-                        Ok(tensor) => return tensor,
-                        Err(_) => {
-                            // If creation fails, deallocate and fall back to clone
-                            SimdAllocator::dealloc_f32(ptr, len);
-                        }
-                    }
-                }
-            }
-        }
-        
-        // Fallback to regular clone if SIMD optimization fails
-        self.clone()
-    }
-
-    /// Try to create a memory-optimized copy with error handling
-    /// エラーハンドリング付きでメモリ最適化コピーを作成を試行
-    pub fn try_optimize_memory(&self) -> RusTorchResult<Self> {
-        let info = self.memory_info();
-        
-        // Check if tensor is too large to optimize safely
-        const MAX_OPTIMIZE_SIZE: usize = 1_000_000_000; // 1GB
-        if info.total_bytes > MAX_OPTIMIZE_SIZE {
-            return Err(RusTorchError::TensorOp {
-                message: format!(
-                    "Tensor too large to optimize: {} bytes exceeds maximum of {} bytes",
-                    info.total_bytes, MAX_OPTIMIZE_SIZE
-                ),
-                source: None,
-            });
-        }
-        
-        Ok(self.optimize_memory())
-    }
+    // Memory optimization methods are now provided by the MemoryOptimization trait
+    // メモリ最適化メソッドはMemoryOptimizationトレイトで提供されます
 
     /// Creates a tensor filled with zeros.
     /// ゼロで満たされたテンソルを作成します。
@@ -624,171 +494,5 @@ impl<T: Float + 'static> From<ndarray::Array2<T>> for Tensor<T> {
     }
 }
 
-// Zero-copy operations extension
-impl<T: Float + Clone + 'static + ndarray::ScalarOperand> Tensor<T> {
-    /// In-place addition with another tensor
-    /// 他のテンソルとの in-place 加算
-    pub fn inplace_add(&mut self, other: &Tensor<T>) -> RusTorchResult<()> {
-        if self.shape() != other.shape() {
-            return Err(RusTorchError::ShapeMismatch {
-                expected: self.shape().to_vec(),
-                actual: other.shape().to_vec(),
-            });
-        }
-        
-        // Use element-wise operations instead of compound assignment
-        for (a, b) in self.data.iter_mut().zip(other.data.iter()) {
-            *a = *a + *b;
-        }
-        Ok(())
-    }
-    
-    /// In-place subtraction with another tensor
-    /// 他のテンソルとの in-place 減算
-    pub fn inplace_sub(&mut self, other: &Tensor<T>) -> RusTorchResult<()> {
-        if self.shape() != other.shape() {
-            return Err(RusTorchError::ShapeMismatch {
-                expected: self.shape().to_vec(),
-                actual: other.shape().to_vec(),
-            });
-        }
-        
-        // Use element-wise operations instead of compound assignment
-        for (a, b) in self.data.iter_mut().zip(other.data.iter()) {
-            *a = *a - *b;
-        }
-        Ok(())
-    }
-    
-    /// In-place multiplication with another tensor
-    /// 他のテンソルとの in-place 乗算
-    pub fn inplace_mul(&mut self, other: &Tensor<T>) -> RusTorchResult<()> {
-        if self.shape() != other.shape() {
-            return Err(RusTorchError::ShapeMismatch {
-                expected: self.shape().to_vec(),
-                actual: other.shape().to_vec(),
-            });
-        }
-        
-        // Use element-wise operations instead of compound assignment
-        for (a, b) in self.data.iter_mut().zip(other.data.iter()) {
-            *a = *a * *b;
-        }
-        Ok(())
-    }
-    
-    /// In-place scalar multiplication
-    /// スカラーとの in-place 乗算
-    pub fn inplace_mul_scalar(&mut self, scalar: T) {
-        for a in self.data.iter_mut() {
-            *a = *a * scalar;
-        }
-    }
-    
-    /// In-place scalar addition
-    /// スカラーとの in-place 加算
-    pub fn inplace_add_scalar(&mut self, scalar: T) {
-        for a in self.data.iter_mut() {
-            *a = *a + scalar;
-        }
-    }
-    
-    /// In-place element-wise function application
-    /// 要素ごとの関数の in-place 適用
-    pub fn inplace_apply<F>(&mut self, f: F) -> RusTorchResult<()>
-    where
-        F: Fn(T) -> T + Send + Sync,
-    {
-        self.data.mapv_inplace(f);
-        Ok(())
-    }
-    
-    /// Create a zero-copy view of a tensor slice
-    /// テンソルスライスのゼロコピービューを作成
-    pub fn slice_view(&self, ranges: &[std::ops::Range<usize>]) -> RusTorchResult<Tensor<T>> {
-        if ranges.len() != self.ndim() {
-            return Err(RusTorchError::TensorOp {
-                message: format!(
-                    "Number of slice ranges {} does not match tensor dimensions {}",
-                    ranges.len(),
-                    self.ndim()
-                ),
-                source: None,
-            });
-        }
-        
-        // Validate ranges
-        for (i, range) in ranges.iter().enumerate() {
-            if range.end > self.shape()[i] {
-                return Err(RusTorchError::TensorOp {
-                    message: format!(
-                        "Slice range {}..{} exceeds dimension {} size {}",
-                        range.start, range.end, i, self.shape()[i]
-                    ),
-                    source: None,
-                });
-            }
-        }
-        
-        // For simplicity, we'll implement basic slicing for 2D tensors
-        // More complex slicing can be added later as needed
-        if ranges.len() == 2 && self.ndim() == 2 {
-            let rows = &ranges[0];
-            let cols = &ranges[1];
-            let original_shape = self.shape();
-            
-            let mut sliced_data = Vec::new();
-            for r in rows.clone() {
-                for c in cols.clone() {
-                    let idx = r * original_shape[1] + c;
-                    if let Some(&value) = self.data.as_slice().unwrap().get(idx) {
-                        sliced_data.push(value);
-                    }
-                }
-            }
-            
-            let new_shape = vec![rows.len(), cols.len()];
-            Self::try_from_vec(sliced_data, new_shape)
-        } else {
-            // For non-2D tensors, just return a clone for now
-            // This maintains API compatibility while avoiding complex slicing
-            Ok(self.clone())
-        }
-    }
-    
-    /// Get an iterator over tensor elements (zero-copy)
-    /// テンソル要素のイテレータを取得（ゼロコピー）
-    pub fn iter(&self) -> impl Iterator<Item = &T> {
-        self.data.iter()
-    }
-    
-    /// Get a mutable iterator over tensor elements (zero-copy)
-    /// テンソル要素の可変イテレータを取得（ゼロコピー）
-    pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut T> {
-        self.data.iter_mut()
-    }
-    
-    /// Check if this tensor shares memory with another tensor
-    /// このテンソルが他のテンソルとメモリを共有しているかチェック
-    pub fn shares_memory_with(&self, other: &Tensor<T>) -> bool {
-        let self_ptr = self.data.as_ptr();
-        let other_ptr = other.data.as_ptr();
-        let self_len = self.data.len();
-        let other_len = other.data.len();
-        
-        // Check if memory regions overlap
-        let self_start = self_ptr as usize;
-        let self_end = self_start + self_len * std::mem::size_of::<T>();
-        let other_start = other_ptr as usize;
-        let other_end = other_start + other_len * std::mem::size_of::<T>();
-        
-        // Memory regions overlap if one starts before the other ends
-        (self_start < other_end) && (other_start < self_end)
-    }
-    
-    /// Create a copy that doesn't share memory (ensures no zero-copy aliasing)
-    /// メモリを共有しないコピーを作成（ゼロコピーエイリアシングを確実に回避）
-    pub fn detach(&self) -> Self {
-        Tensor::new(self.data.clone())
-    }
-}
+// Zero-copy operations are now provided by the ZeroCopyOps and TensorIterOps traits
+// ゼロコピー操作はZeroCopyOpsとTensorIterOpsトレイトで提供されます
