+++ src/tensor/memory/aligned.rs
