+++ src/tensor/memory/pool.rs
@@ -1,0 +1,6 @@
+//! Advanced memory pooling strategies for tensor allocation
+//! テンソル割り当てのための高度なメモリプーリング戦略
+
+// Re-export from the main memory module for now
+// We can move the actual implementation here later
+pub use crate::memory::*;
