--- src/tensor/operations.rs
@@ -1,2729 +1,0 @@
-//! Mathematical operations for tensors
-//! テンソルの数学演算
-
-use super::core::Tensor;
-// Removed unused imports
-use num_complex::Complex;
-use num_traits::Float;
-use std::ops;
-
-impl<T: Float + 'static + ndarray::ScalarOperand + num_traits::FromPrimitive> Tensor<T> {
-    /// Create tensor with random values from normal distribution
-    /// 正規分布からランダム値でテンソルを作成
-    pub fn randn(shape: &[usize]) -> Tensor<T>
-    where
-        T: From<f32>,
-    {
-        use rand::prelude::*;
-        use rand_distr::StandardNormal;
-
-        let mut rng = thread_rng();
-        let total_size: usize = shape.iter().product();
-        let data: Vec<T> = (0..total_size)
-            .map(|_| <T as From<f32>>::from(rng.sample::<f32, _>(StandardNormal)))
-            .collect();
-
-        Tensor::from_vec(data, shape.to_vec())
-    }
-
-    /// Element-wise maximum with another tensor
-    /// 別のテンソルとの要素ごとの最大値
-    pub fn maximum(&self, other: &Tensor<T>) -> Result<Self, String> {
-        if self.shape() != other.shape() {
-            return Err("Shape mismatch for maximum operation".to_string());
-        }
-
-        let max_data: Vec<T> = self
-            .data
-            .iter()
-            .zip(other.data.iter())
-            .map(|(&a, &b)| if a > b { a } else { b })
-            .collect();
-
-        Ok(Tensor::from_vec(max_data, self.shape().to_vec()))
-    }
-
-    /// Stack tensors along a new axis
-    /// 新しい軸に沿ってテンソルをスタック
-    pub fn stack(tensors: &[&Tensor<T>]) -> Result<Tensor<T>, String> {
-        if tensors.is_empty() {
-            return Err("Cannot stack empty tensor list".to_string());
-        }
-
-        let first_shape = tensors[0].shape();
-        for tensor in tensors.iter().skip(1) {
-            if tensor.shape() != first_shape {
-                return Err("All tensors must have the same shape".to_string());
-            }
-        }
-
-        // Create new shape with additional dimension
-        let mut new_shape = vec![tensors.len()];
-        new_shape.extend_from_slice(first_shape);
-
-        // Collect all data
-        let mut all_data = Vec::new();
-        for tensor in tensors {
-            all_data.extend_from_slice(tensor.data.as_slice().unwrap_or(&[]));
-        }
-
-        Ok(Tensor::from_vec(all_data, new_shape))
-    }
-    /// Element-wise addition with another tensor (supports broadcasting).
-    /// 別のテンソルとの要素ごとの加算（ブロードキャスト対応）
-    pub fn add(&self, other: &Tensor<T>) -> Result<Self, String> {
-        // Check if shapes are compatible for broadcasting
-        if self.shape() == other.shape() {
-            // Direct addition when shapes match
-            let result_data = self.as_array() + other.as_array();
-            Ok(Tensor::new(result_data))
-        } else if self.can_broadcast_with(other) {
-            // Simple broadcasting for common case: (N, M) + (1, M)
-            let self_shape = self.shape();
-            let other_shape = other.shape();
-
-            // Case 1: (N, M) + (1, M) - bias addition pattern
-            if self_shape.len() == 2
-                && other_shape.len() == 2
-                && self_shape[1] == other_shape[1]
-                && other_shape[0] == 1
-            {
-                let batch_size = self_shape[0];
-                let feature_size = self_shape[1];
-                let mut result_data = vec![T::zero(); batch_size * feature_size];
-
-                for b in 0..batch_size {
-                    for f in 0..feature_size {
-                        let self_idx = b * feature_size + f;
-                        let other_idx = f; // broadcast from (1, M)
-                        result_data[self_idx] = self.data.as_slice().unwrap()[self_idx]
-                            + other.data.as_slice().unwrap()[other_idx];
-                    }
-                }
-
-                Ok(Tensor::from_vec(result_data, self_shape.to_vec()))
-            }
-            // Case 2: (N, M) + [1] - scalar broadcast
-            else if other_shape.len() == 1 && other_shape[0] == 1 {
-                let scalar_val = other.data.as_slice().unwrap()[0];
-                let result_data: Vec<T> = self
-                    .data
-                    .as_slice()
-                    .unwrap()
-                    .iter()
-                    .map(|&x| x + scalar_val)
-                    .collect();
-                Ok(Tensor::from_vec(result_data, self_shape.to_vec()))
-            }
-            // Case 3: [1] + (N, M) - scalar broadcast (commutative)
-            else if self_shape.len() == 1 && self_shape[0] == 1 {
-                let scalar_val = self.data.as_slice().unwrap()[0];
-                let result_data: Vec<T> = other
-                    .data
-                    .as_slice()
-                    .unwrap()
-                    .iter()
-                    .map(|&x| scalar_val + x)
-                    .collect();
-                Ok(Tensor::from_vec(result_data, other_shape.to_vec()))
-            }
-            // Case 4: 4D tensor broadcasting for batch norm: (N, C, H, W) + (1, C, 1, 1)
-            else if self_shape.len() == 4
-                && other_shape.len() == 4
-                && self_shape[1] == other_shape[1]
-                && other_shape[0] == 1
-                && other_shape[2] == 1
-                && other_shape[3] == 1
-            {
-                let (n, c, h, w) = (self_shape[0], self_shape[1], self_shape[2], self_shape[3]);
-                let mut result_data = vec![T::zero(); n * c * h * w];
-
-                for batch in 0..n {
-                    for ch in 0..c {
-                        let bias_val = other.data.as_slice().unwrap()[ch];
-                        for height in 0..h {
-                            for width in 0..w {
-                                let idx = batch * (c * h * w) + ch * (h * w) + height * w + width;
-                                result_data[idx] = self.data.as_slice().unwrap()[idx] + bias_val;
-                            }
-                        }
-                    }
-                }
-                Ok(Tensor::from_vec(result_data, self_shape.to_vec()))
-            } else {
-                Err(format!(
-                    "Broadcasting not implemented for shapes: {:?} vs {:?}",
-                    self_shape, other_shape
-                ))
-            }
-        } else {
-            Err(format!(
-                "Shape mismatch: {:?} vs {:?}",
-                self.shape(),
-                other.shape()
-            ))
-        }
-    }
-
-    /// Element-wise subtraction with another tensor.
-    /// 別のテンソルとの要素ごとの減算
-    pub fn sub(&self, other: &Tensor<T>) -> Result<Self, String> {
-        if self.shape() != other.shape() {
-            return Err(format!(
-                "Shape mismatch: {:?} vs {:?}",
-                self.shape(),
-                other.shape()
-            ));
-        }
-
-        let result_data = self.as_array() - other.as_array();
-        Ok(Tensor::new(result_data))
-    }
-
-    /// Element-wise multiplication with another tensor (supports broadcasting).
-    /// 別のテンソルとの要素ごとの乗算（ブロードキャスト対応）
-    pub fn mul(&self, other: &Tensor<T>) -> Result<Self, String> {
-        // Check if shapes are compatible for broadcasting
-        if self.shape() == other.shape() {
-            // Direct multiplication when shapes match
-            let result_data = self.as_array() * other.as_array();
-            Ok(Tensor::new(result_data))
-        } else if self.can_broadcast_with(other) {
-            // Broadcasting for multiplication
-            let self_shape = self.shape();
-            let other_shape = other.shape();
-
-            // Case 1: (N, M) * (1, M) - bias multiplication pattern
-            if self_shape.len() == 2
-                && other_shape.len() == 2
-                && self_shape[1] == other_shape[1]
-                && other_shape[0] == 1
-            {
-                let batch_size = self_shape[0];
-                let feature_size = self_shape[1];
-                let mut result_data = vec![T::zero(); batch_size * feature_size];
-
-                for b in 0..batch_size {
-                    for f in 0..feature_size {
-                        let self_idx = b * feature_size + f;
-                        let other_idx = f; // broadcast from (1, M)
-                        result_data[self_idx] = self.data.as_slice().unwrap()[self_idx]
-                            * other.data.as_slice().unwrap()[other_idx];
-                    }
-                }
-
-                Ok(Tensor::from_vec(result_data, self_shape.to_vec()))
-            }
-            // Case 2: (N, M) * [1] - scalar broadcast
-            else if other_shape.len() == 1 && other_shape[0] == 1 {
-                let scalar_val = other.data.as_slice().unwrap()[0];
-                let result_data: Vec<T> = self
-                    .data
-                    .as_slice()
-                    .unwrap()
-                    .iter()
-                    .map(|&x| x * scalar_val)
-                    .collect();
-                Ok(Tensor::from_vec(result_data, self_shape.to_vec()))
-            }
-            // Case 3: [1] * (N, M) - scalar broadcast (commutative)
-            else if self_shape.len() == 1 && self_shape[0] == 1 {
-                let scalar_val = self.data.as_slice().unwrap()[0];
-                let result_data: Vec<T> = other
-                    .data
-                    .as_slice()
-                    .unwrap()
-                    .iter()
-                    .map(|&x| scalar_val * x)
-                    .collect();
-                Ok(Tensor::from_vec(result_data, other_shape.to_vec()))
-            } else {
-                Err(format!(
-                    "Broadcasting not implemented for shapes: {:?} vs {:?}",
-                    self_shape, other_shape
-                ))
-            }
-        } else {
-            Err(format!(
-                "Shape mismatch: {:?} vs {:?}",
-                self.shape(),
-                other.shape()
-            ))
-        }
-    }
-
-    /// Element-wise division with another tensor.
-    /// 別のテンソルとの要素ごとの除算
-    pub fn div(&self, other: &Tensor<T>) -> Result<Self, String> {
-        if self.shape() != other.shape() {
-            return Err(format!(
-                "Shape mismatch: {:?} vs {:?}",
-                self.shape(),
-                other.shape()
-            ));
-        }
-
-        let result_data = self.as_array() / other.as_array();
-        Ok(Tensor::new(result_data))
-    }
-
-    /// Scalar addition.
-    /// スカラー加算
-    pub fn add_scalar(&self, scalar: T) -> Self {
-        let result_data = self.as_array() + scalar;
-        Tensor::new(result_data)
-    }
-
-    /// Scalar subtraction.
-    /// スカラー減算
-    pub fn sub_scalar(&self, scalar: T) -> Self {
-        let result_data = self.as_array() - scalar;
-        Tensor::new(result_data)
-    }
-
-    /// Scalar multiplication.
-    /// スカラー乗算
-    pub fn mul_scalar(&self, scalar: T) -> Self {
-        let result_data = self.as_array() * scalar;
-        Tensor::new(result_data)
-    }
-
-    /// Scalar division.
-    /// スカラー除算
-    pub fn div_scalar(&self, scalar: T) -> Self {
-        let result_data = self.as_array() / scalar;
-        Tensor::new(result_data)
-    }
-
-    /// Negation of all elements.
-    /// 全要素の符号反転
-    pub fn neg(&self) -> Self {
-        let result_data = self.as_array().mapv(|x| -x);
-        Tensor::new(result_data)
-    }
-
-    /// Matrix multiplication.
-    /// 行列乗算
-    pub fn matmul(&self, other: &Tensor<T>) -> Result<Self, String> {
-        // Support 2D x 2D, 3D x 2D, and 4D x 4D operations
-        if !(self.ndim() == 2 && other.ndim() == 2)
-            && !(self.ndim() == 3 && other.ndim() == 2)
-            && !(self.ndim() == 4 && other.ndim() == 4)
-        {
-            return Err(format!(
-                "Unsupported matmul dimensions: {}D @ {}D",
-                self.ndim(),
-                other.ndim()
-            ));
-        }
-
-        let self_shape = self.shape();
-        let other_shape = other.shape();
-
-        // Handle 2D x 2D case
-        if self.ndim() == 2 && other.ndim() == 2 {
-            if self_shape[1] != other_shape[0] {
-                return Err(format!(
-                    "Matrix dimension mismatch: {}x{} @ {}x{}",
-                    self_shape[0], self_shape[1], other_shape[0], other_shape[1]
-                ));
-            }
-
-            let result_shape = vec![self_shape[0], other_shape[1]];
-            let mut result_data = vec![T::zero(); result_shape.iter().product()];
-
-            let self_array = &self.data;
-            let other_array = &other.data;
-
-            for i in 0..result_shape[0] {
-                for j in 0..result_shape[1] {
-                    let mut sum = T::zero();
-                    for k in 0..self_shape[1] {
-                        let self_val = self_array[[i, k]];
-                        let other_val = other_array[[k, j]];
-                        sum = sum + self_val * other_val;
-                    }
-                    result_data[i * result_shape[1] + j] = sum;
-                }
-            }
-
-            Ok(Tensor::from_vec(result_data, result_shape))
-        }
-        // Handle 3D x 2D (batch) case
-        else if self.ndim() == 3 && other.ndim() == 2 {
-            let batch_size = self_shape[0];
-            let seq_len = self_shape[1];
-            let input_feat = self_shape[2];
-            let output_feat = other_shape[1];
-
-            if input_feat != other_shape[0] {
-                return Err(format!(
-                    "Feature dimension mismatch: {} != {}",
-                    input_feat, other_shape[0]
-                ));
-            }
-
-            let result_shape = vec![batch_size, seq_len, output_feat];
-            let mut result_data = vec![T::zero(); result_shape.iter().product()];
-
-            for b in 0..batch_size {
-                for s in 0..seq_len {
-                    for o in 0..output_feat {
-                        let mut sum = T::zero();
-                        for i in 0..input_feat {
-                            let self_val = self.data[[b, s, i]];
-                            let other_val = other.data[[i, o]];
-                            sum = sum + self_val * other_val;
-                        }
-                        result_data[b * seq_len * output_feat + s * output_feat + o] = sum;
-                    }
-                }
-            }
-
-            Ok(Tensor::from_vec(result_data, result_shape))
-        }
-        // Handle 4D x 4D (batch, heads, seq, features) case for attention
-        else if self.ndim() == 4 && other.ndim() == 4 {
-            let batch_size = self_shape[0];
-            let num_heads = self_shape[1];
-            let seq_len_q = self_shape[2];
-            let feat_q = self_shape[3];
-
-            // For attention: Q @ K^T, where K^T has shape (batch, heads, features, seq_len)
-            // So we expect other to be transposed: (batch, heads, feat_k, seq_len_k)
-            let feat_k = other_shape[2];
-            let seq_len_k = other_shape[3];
-
-            if feat_q != feat_k {
-                return Err(format!(
-                    "Feature dimension mismatch: {} != {}",
-                    feat_q, feat_k
-                ));
-            }
-
-            let result_shape = vec![batch_size, num_heads, seq_len_q, seq_len_k];
-            let mut result_data = vec![T::zero(); result_shape.iter().product()];
-
-            for b in 0..batch_size {
-                for h in 0..num_heads {
-                    for i in 0..seq_len_q {
-                        for j in 0..seq_len_k {
-                            let mut sum = T::zero();
-                            for k in 0..feat_q {
-                                let self_val = self.data[[b, h, i, k]];
-                                let other_val = other.data[[b, h, k, j]]; // Transposed access
-                                sum = sum + self_val * other_val;
-                            }
-                            let idx = b * num_heads * seq_len_q * seq_len_k
-                                + h * seq_len_q * seq_len_k
-                                + i * seq_len_k
-                                + j;
-                            result_data[idx] = sum;
-                        }
-                    }
-                }
-            }
-
-            Ok(Tensor::from_vec(result_data, result_shape))
-        } else {
-            Err("Unsupported tensor dimensions for matmul".to_string())
-        }
-    }
-
-    /// Transpose the tensor (swap last two dimensions).
-    /// テンソルの転置（最後の2次元を入れ替え）
-    pub fn transpose(&self) -> Result<Self, String> {
-        if self.ndim() < 2 {
-            return Err("Transpose requires at least 2D tensor".to_string());
-        }
-
-        let mut axes: Vec<usize> = (0..self.ndim()).collect();
-        let ndim = self.ndim();
-        axes.swap(ndim - 2, ndim - 1);
-
-        let transposed = self
-            .as_array()
-            .clone()
-            .permuted_axes(axes.into_iter().collect::<Vec<_>>());
-        Ok(Tensor::new(transposed))
-    }
-
-    /// Sum of all elements.
-    /// 全要素の合計
-    pub fn sum(&self) -> T {
-        self.as_array().sum()
-    }
-
-    /// Mean of all elements.
-    /// 全要素の平均
-    pub fn mean(&self) -> T {
-        let sum = self.sum();
-        let count = T::from(self.numel()).unwrap_or(T::one());
-        sum / count
-    }
-
-    /// Singular Value Decomposition (SVD)
-    /// 特異値分解 - torch.svd compatible
-    ///
-    /// Decomposes a matrix A into U, S, V^T such that A = U * S * V^T
-    /// Returns (U, S, V) where:
-    /// - U: left singular vectors (m x min(m,n))
-    /// - S: singular values (min(m,n),)  
-    /// - V: right singular vectors (n x min(m,n))
-    pub fn svd(&self, some: bool) -> Result<(Self, Self, Self), String> {
-        if self.ndim() < 2 {
-            return Err("SVD requires at least 2D tensor".to_string());
-        }
-
-        // For now, handle 2D case (can be extended to batched SVD later)
-        if self.ndim() != 2 {
-            return Err("Batched SVD not yet implemented - use 2D tensors".to_string());
-        }
-
-        let shape = self.shape();
-        let m = shape[0]; // rows
-        let n = shape[1]; // cols
-        let min_mn = m.min(n);
-
-        // Convert to ndarray for SVD computation
-        let _matrix = self.as_array();
-
-        // Compute SVD using ndarray's linear algebra (if available)
-        // For now, implement a simplified version
-        self.svd_impl(m, n, min_mn, some)
-    }
-
-    /// Internal SVD implementation
-    fn svd_impl(
-        &self,
-        m: usize,
-        n: usize,
-        min_mn: usize,
-        some: bool,
-    ) -> Result<(Self, Self, Self), String> {
-        // Implementation using power iteration method for educational purposes
-        // In production, use LAPACK bindings for optimal performance
-
-        #[cfg(all(
-            any(feature = "linalg", feature = "linalg-system"),
-            not(target_arch = "wasm32")
-        ))]
-        {
-            return self.svd_with_linalg(m, n, min_mn, some);
-        }
-
-        #[cfg(not(all(
-            any(feature = "linalg", feature = "linalg-system"),
-            not(target_arch = "wasm32")
-        )))]
-        {
-            return self.svd_basic(m, n, min_mn, some);
-        }
-    }
-
-    /// Basic SVD implementation without external linear algebra library
-    fn svd_basic(
-        &self,
-        m: usize,
-        n: usize,
-        min_mn: usize,
-        _some: bool,
-    ) -> Result<(Self, Self, Self), String> {
-        let matrix = self.as_array();
-
-        // Create A^T * A manually to avoid type issues
-        // Compute singular values using approximation based on column norms
-        let mut s_data = vec![T::zero(); min_mn];
-
-        // Simple approximation: compute column norms as singular values
-        for j in 0..min_mn {
-            let mut col_norm_sq = T::zero();
-            for i in 0..m {
-                if j < n {
-                    let val = matrix[[i, j]];
-                    col_norm_sq = col_norm_sq + val * val;
-                }
-            }
-            s_data[j] = col_norm_sq.sqrt();
-        }
-
-        // Sort singular values in descending order
-        s_data.sort_by(|a, b| b.partial_cmp(a).unwrap_or(std::cmp::Ordering::Equal));
-        let s = Tensor::from_vec(s_data.clone(), vec![min_mn]);
-
-        // Compute U and V using Gram-Schmidt process (simplified)
-        // For basic implementation, always use reduced form regardless of 'some' parameter
-        let u_size = min_mn;
-        let v_size = min_mn;
-
-        // Initialize U with first few columns of A (normalized)
-        let mut u_data = vec![T::zero(); m * u_size];
-        for j in 0..u_size.min(n) {
-            let mut col_norm = T::zero();
-            for i in 0..m {
-                let val = matrix[[i, j]];
-                u_data[i * u_size + j] = val;
-                col_norm = col_norm + val * val;
-            }
-            col_norm = col_norm.sqrt();
-
-            if col_norm > T::from(1e-10).unwrap_or(T::zero()) {
-                for i in 0..m {
-                    u_data[i * u_size + j] = u_data[i * u_size + j] / col_norm;
-                }
-            }
-        }
-        let u = Tensor::from_vec(u_data, vec![m, u_size]);
-
-        // Initialize V with first few rows of A^T (normalized)
-        let mut v_data = vec![T::zero(); n * v_size];
-        for j in 0..v_size.min(m) {
-            let mut row_norm = T::zero();
-            for i in 0..n {
-                let val = matrix[[j, i]];
-                v_data[i * v_size + j] = val;
-                row_norm = row_norm + val * val;
-            }
-            row_norm = row_norm.sqrt();
-
-            if row_norm > T::from(1e-10).unwrap_or(T::zero()) {
-                for i in 0..n {
-                    v_data[i * v_size + j] = v_data[i * v_size + j] / row_norm;
-                }
-            }
-        }
-        let v = Tensor::from_vec(v_data, vec![n, v_size]);
-
-        Ok((u, s, v))
-    }
-
-    /// SVD implementation with ndarray-linalg (more accurate)
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn svd_with_linalg(
-        &self,
-        m: usize,
-        n: usize,
-        min_mn: usize,
-        some: bool,
-    ) -> Result<(Self, Self, Self), String> {
-        use ndarray_linalg::SVD;
-
-        let matrix = self.as_array().clone();
-
-        // Convert dynamic array to 2D array for linalg operations
-        let matrix_2d = matrix
-            .into_dimensionality::<ndarray::Ix2>()
-            .map_err(|e| format!("Failed to convert to 2D array: {:?}", e))?;
-
-        // For f32, convert to f64 for computation, then back to f32
-        if std::any::TypeId::of::<T>() == std::any::TypeId::of::<f32>() {
-            // Convert f32 matrix to f64
-            let matrix_f64 = matrix_2d.mapv(|x| x.to_f64().unwrap_or(0.0));
-
-            let (u, s, vt) = matrix_f64
-                .svd(true, true)
-                .map_err(|e| format!("SVD computation failed: {:?}", e))?;
-
-            let u = u.unwrap();
-            let vt = vt.unwrap();
-
-            // Convert back to tensor type
-            let u_data: Vec<T> = u.iter().map(|&x| T::from(x).unwrap_or(T::zero())).collect();
-            let s_data: Vec<T> = s.iter().map(|&x| T::from(x).unwrap_or(T::zero())).collect();
-            let v_data: Vec<T> = vt
-                .t()
-                .iter()
-                .map(|&x| T::from(x).unwrap_or(T::zero()))
-                .collect();
-
-            let u_tensor = Tensor::from_vec(u_data, vec![u.nrows(), u.ncols()]);
-            let s_tensor = Tensor::from_vec(s_data, vec![s.len()]);
-            let v_tensor = Tensor::from_vec(v_data, vec![vt.ncols(), vt.nrows()]);
-
-            Ok((u_tensor, s_tensor, v_tensor))
-        } else if std::any::TypeId::of::<T>() == std::any::TypeId::of::<f64>() {
-            // For f64, use directly
-            let matrix_f64 = matrix_2d.mapv(|x| x.to_f64().unwrap_or(0.0));
-
-            let (u, s, vt) = matrix_f64
-                .svd(true, true)
-                .map_err(|e| format!("SVD computation failed: {:?}", e))?;
-
-            let u = u.unwrap();
-            let vt = vt.unwrap();
-
-            let u_data: Vec<T> = u.iter().map(|&x| T::from(x).unwrap_or(T::zero())).collect();
-            let s_data: Vec<T> = s.iter().map(|&x| T::from(x).unwrap_or(T::zero())).collect();
-            let v_data: Vec<T> = vt
-                .t()
-                .iter()
-                .map(|&x| T::from(x).unwrap_or(T::zero()))
-                .collect();
-
-            let u_tensor = Tensor::from_vec(u_data, vec![u.nrows(), u.ncols()]);
-            let s_tensor = Tensor::from_vec(s_data, vec![s.len()]);
-            let v_tensor = Tensor::from_vec(v_data, vec![vt.ncols(), vt.nrows()]);
-
-            Ok((u_tensor, s_tensor, v_tensor))
-        } else {
-            // Fallback to basic implementation for other types
-            self.svd_basic(m, n, min_mn, some)
-        }
-    }
-
-    /// Eigenvalue decomposition for general matrices - torch.eig compatible
-    /// 一般行列の固有値分解 - torch.eig互換
-    ///
-    /// Returns (eigenvalues, eigenvectors) where:
-    /// - eigenvalues: complex eigenvalues (real_part + i * imag_part) as [n, 2] tensor
-    /// - eigenvectors: right eigenvectors as [n, n] tensor (if eigenvectors=true)
-    pub fn eig(&self, eigenvectors: bool) -> Result<(Self, Option<Self>), String> {
-        if self.ndim() != 2 {
-            return Err("eig() only supports 2D tensors".to_string());
-        }
-
-        let shape = self.shape();
-        if shape[0] != shape[1] {
-            return Err("eig() only supports square matrices".to_string());
-        }
-
-        let n = shape[0];
-
-        #[cfg(all(
-            any(feature = "linalg", feature = "linalg-system"),
-            not(target_arch = "wasm32")
-        ))]
-        {
-            return self.eig_with_linalg(n, eigenvectors);
-        }
-
-        #[cfg(not(all(
-            any(feature = "linalg", feature = "linalg-system"),
-            not(target_arch = "wasm32")
-        )))]
-        {
-            return self.eig_basic(n, eigenvectors);
-        }
-    }
-
-    /// Symmetric eigenvalue decomposition - torch.symeig compatible
-    /// 対称行列の固有値分解 - torch.symeig互換
-    ///
-    /// Returns (eigenvalues, eigenvectors) where:
-    /// - eigenvalues: real eigenvalues as [n] tensor (sorted in ascending order if upper=false)
-    /// - eigenvectors: orthonormal eigenvectors as [n, n] tensor (if eigenvectors=true)
-    pub fn symeig(&self, eigenvectors: bool, upper: bool) -> Result<(Self, Option<Self>), String> {
-        if self.ndim() != 2 {
-            return Err("symeig() only supports 2D tensors".to_string());
-        }
-
-        let shape = self.shape();
-        if shape[0] != shape[1] {
-            return Err("symeig() only supports square matrices".to_string());
-        }
-
-        let n = shape[0];
-
-        #[cfg(all(
-            any(feature = "linalg", feature = "linalg-system"),
-            not(target_arch = "wasm32")
-        ))]
-        {
-            return self.symeig_with_linalg(n, eigenvectors, upper);
-        }
-
-        #[cfg(not(all(
-            any(feature = "linalg", feature = "linalg-system"),
-            not(target_arch = "wasm32")
-        )))]
-        {
-            return self.symeig_basic(n, eigenvectors, upper);
-        }
-    }
-
-    /// Basic eigenvalue decomposition implementation
-    pub fn eig_basic(&self, n: usize, eigenvectors: bool) -> Result<(Self, Option<Self>), String> {
-        // Basic implementation using power iteration method
-        // This is a simplified implementation for educational purposes
-
-        let matrix = self.as_array();
-
-        // For basic implementation, compute approximate eigenvalues using diagonal elements
-        let mut eigenvals_real = vec![T::zero(); n];
-        let eigenvals_imag = vec![T::zero(); n];
-
-        // Simple approximation: use diagonal elements as eigenvalue estimates
-        for i in 0..n {
-            eigenvals_real[i] = matrix[[i, i]];
-            // Imaginary parts are zero for this basic approximation
-        }
-
-        // Create eigenvalues tensor as [n, 2] (real, imag)
-        let mut eigenvals_data = Vec::with_capacity(n * 2);
-        for i in 0..n {
-            eigenvals_data.push(eigenvals_real[i]);
-            eigenvals_data.push(eigenvals_imag[i]);
-        }
-        let eigenvals = Tensor::from_vec(eigenvals_data, vec![n, 2]);
-
-        let eigenvecs = if eigenvectors {
-            // Return identity matrix as approximate eigenvectors
-            let mut eigenvec_data = vec![T::zero(); n * n];
-            for i in 0..n {
-                eigenvec_data[i * n + i] = T::one();
-            }
-            Some(Tensor::from_vec(eigenvec_data, vec![n, n]))
-        } else {
-            None
-        };
-
-        Ok((eigenvals, eigenvecs))
-    }
-
-    /// Basic symmetric eigenvalue decomposition implementation  
-    pub fn symeig_basic(
-        &self,
-        n: usize,
-        eigenvectors: bool,
-        _upper: bool,
-    ) -> Result<(Self, Option<Self>), String> {
-        // Basic implementation for symmetric matrices
-        let matrix = self.as_array();
-
-        // Simple approximation: use diagonal elements as eigenvalues
-        let mut eigenvals_data = vec![T::zero(); n];
-        for i in 0..n {
-            eigenvals_data[i] = matrix[[i, i]];
-        }
-
-        // Sort eigenvalues (ascending order by default)
-        eigenvals_data.sort_by(|a, b| a.partial_cmp(b).unwrap_or(std::cmp::Ordering::Equal));
-
-        let eigenvals = Tensor::from_vec(eigenvals_data, vec![n]);
-
-        let eigenvecs = if eigenvectors {
-            // Return identity matrix as approximate eigenvectors
-            let mut eigenvec_data = vec![T::zero(); n * n];
-            for i in 0..n {
-                eigenvec_data[i * n + i] = T::one();
-            }
-            Some(Tensor::from_vec(eigenvec_data, vec![n, n]))
-        } else {
-            None
-        };
-
-        Ok((eigenvals, eigenvecs))
-    }
-
-    /// Eigenvalue decomposition using ndarray-linalg
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn eig_with_linalg(
-        &self,
-        n: usize,
-        eigenvectors: bool,
-    ) -> Result<(Self, Option<Self>), String> {
-        use ndarray_linalg::Eig;
-
-        let matrix = self.as_array().clone();
-
-        // Convert to 2D array for linalg operations
-        let matrix_2d = matrix
-            .into_dimensionality::<ndarray::Ix2>()
-            .map_err(|e| format!("Failed to convert to 2D array: {:?}", e))?;
-
-        // Convert to f64 for computation
-        let matrix_f64 = matrix_2d.mapv(|x| x.to_f64().unwrap_or(0.0));
-
-        if eigenvectors {
-            let (eigenvals, eigenvecs) = matrix_f64
-                .eig()
-                .map_err(|e| format!("Eigenvalue computation failed: {:?}", e))?;
-
-            // Convert eigenvalues from complex to [n, 2] format (real, imag)
-            let mut eigenvals_data = Vec::with_capacity(n * 2);
-            for val in eigenvals.iter() {
-                eigenvals_data.push(T::from(val.re).unwrap_or(T::zero()));
-                eigenvals_data.push(T::from(val.im).unwrap_or(T::zero()));
-            }
-            let eigenvals_tensor = Tensor::from_vec(eigenvals_data, vec![n, 2]);
-
-            // Convert eigenvectors back to tensor type
-            let eigenvec_data: Vec<T> = eigenvecs
-                .iter()
-                .map(|&x| T::from(x.re).unwrap_or(T::zero()))
-                .collect();
-            let eigenvecs_tensor = Tensor::from_vec(eigenvec_data, vec![n, n]);
-
-            Ok((eigenvals_tensor, Some(eigenvecs_tensor)))
-        } else {
-            let eigenvals = matrix_f64
-                .eig()
-                .map_err(|e| format!("Eigenvalue computation failed: {:?}", e))?
-                .0;
-
-            // Convert eigenvalues from complex to [n, 2] format
-            let mut eigenvals_data = Vec::with_capacity(n * 2);
-            for val in eigenvals.iter() {
-                eigenvals_data.push(T::from(val.re).unwrap_or(T::zero()));
-                eigenvals_data.push(T::from(val.im).unwrap_or(T::zero()));
-            }
-            let eigenvals_tensor = Tensor::from_vec(eigenvals_data, vec![n, 2]);
-
-            Ok((eigenvals_tensor, None))
-        }
-    }
-
-    /// Symmetric eigenvalue decomposition using ndarray-linalg
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn symeig_with_linalg(
-        &self,
-        n: usize,
-        eigenvectors: bool,
-        upper: bool,
-    ) -> Result<(Self, Option<Self>), String> {
-        use ndarray_linalg::Eigh;
-        use ndarray_linalg::UPLO;
-
-        let matrix = self.as_array().clone();
-
-        // Convert to 2D array for linalg operations
-        let matrix_2d = matrix
-            .into_dimensionality::<ndarray::Ix2>()
-            .map_err(|e| format!("Failed to convert to 2D array: {:?}", e))?;
-
-        // Convert to f64 for computation
-        let matrix_f64 = matrix_2d.mapv(|x| x.to_f64().unwrap_or(0.0));
-
-        let uplo = if upper { UPLO::Upper } else { UPLO::Lower };
-
-        if eigenvectors {
-            let (eigenvals, eigenvecs) = matrix_f64
-                .eigh(uplo)
-                .map_err(|e| format!("Symmetric eigenvalue computation failed: {:?}", e))?;
-
-            // Convert eigenvalues back to tensor type
-            let eigenvals_data: Vec<T> = eigenvals
-                .iter()
-                .map(|&x| T::from(x).unwrap_or(T::zero()))
-                .collect();
-            let eigenvals_tensor = Tensor::from_vec(eigenvals_data, vec![n]);
-
-            // Convert eigenvectors back to tensor type
-            let eigenvec_data: Vec<T> = eigenvecs
-                .iter()
-                .map(|&x| T::from(x).unwrap_or(T::zero()))
-                .collect();
-            let eigenvecs_tensor = Tensor::from_vec(eigenvec_data, vec![n, n]);
-
-            Ok((eigenvals_tensor, Some(eigenvecs_tensor)))
-        } else {
-            let eigenvals = matrix_f64
-                .eigh(uplo)
-                .map_err(|e| format!("Symmetric eigenvalue computation failed: {:?}", e))?
-                .0;
-
-            // Convert eigenvalues back to tensor type
-            let eigenvals_data: Vec<T> = eigenvals
-                .iter()
-                .map(|&x| T::from(x).unwrap_or(T::zero()))
-                .collect();
-            let eigenvals_tensor = Tensor::from_vec(eigenvals_data, vec![n]);
-
-            Ok((eigenvals_tensor, None))
-        }
-    }
-
-    /// QR decomposition - A = Q * R
-    /// QR分解 - A = Q * R
-    ///
-    /// Returns (Q, R) where:
-    /// - Q: orthogonal matrix (m x min(m,n))
-    /// - R: upper triangular matrix (min(m,n) x n)
-    pub fn qr(&self) -> Result<(Self, Self), String> {
-        if self.ndim() != 2 {
-            return Err("QR decomposition only supports 2D tensors".to_string());
-        }
-
-        let shape = self.shape();
-        let m = shape[0]; // rows
-        let n = shape[1]; // cols
-        let min_mn = m.min(n);
-
-        #[cfg(all(
-            any(feature = "linalg", feature = "linalg-system"),
-            not(target_arch = "wasm32")
-        ))]
-        {
-            return self.qr_with_linalg(m, n, min_mn);
-        }
-
-        #[cfg(not(all(
-            any(feature = "linalg", feature = "linalg-system"),
-            not(target_arch = "wasm32")
-        )))]
-        {
-            return self.qr_basic(m, n, min_mn);
-        }
-    }
-
-    /// LU decomposition with partial pivoting - PA = LU  
-    /// LU分解（部分ピボット付き）- PA = LU
-    ///
-    /// Returns (L, U, P) where:
-    /// - L: lower triangular matrix with unit diagonal (m x min(m,n))
-    /// - U: upper triangular matrix (min(m,n) x n)
-    /// - P: permutation matrix (m x m)
-    pub fn lu(&self) -> Result<(Self, Self, Self), String> {
-        if self.ndim() != 2 {
-            return Err("LU decomposition only supports 2D tensors".to_string());
-        }
-
-        let shape = self.shape();
-        let m = shape[0]; // rows
-        let n = shape[1]; // cols
-        let min_mn = m.min(n);
-
-        #[cfg(all(
-            any(feature = "linalg", feature = "linalg-system"),
-            not(target_arch = "wasm32")
-        ))]
-        {
-            return self.lu_with_linalg(m, n, min_mn);
-        }
-
-        #[cfg(not(all(
-            any(feature = "linalg", feature = "linalg-system"),
-            not(target_arch = "wasm32")
-        )))]
-        {
-            return self.lu_basic(m, n, min_mn);
-        }
-    }
-
-    /// Basic QR decomposition using Gram-Schmidt process
-    pub fn qr_basic(&self, m: usize, n: usize, min_mn: usize) -> Result<(Self, Self), String> {
-        let matrix = self.as_array();
-
-        // Initialize Q and R matrices
-        let mut q_data = vec![T::zero(); m * min_mn];
-        let mut r_data = vec![T::zero(); min_mn * n];
-
-        // Gram-Schmidt process
-        for j in 0..n.min(min_mn) {
-            // Copy column j of A to column j of Q
-            for i in 0..m {
-                q_data[i * min_mn + j] = matrix[[i, j]];
-            }
-
-            // Orthogonalize against previous columns
-            for k in 0..j {
-                // Compute dot product of current column with column k of Q
-                let mut dot_product = T::zero();
-                for i in 0..m {
-                    dot_product = dot_product + q_data[i * min_mn + j] * q_data[i * min_mn + k];
-                }
-
-                // Store in R matrix
-                r_data[k * n + j] = dot_product;
-
-                // Subtract projection
-                for i in 0..m {
-                    q_data[i * min_mn + j] =
-                        q_data[i * min_mn + j] - dot_product * q_data[i * min_mn + k];
-                }
-            }
-
-            // Normalize the column (compute norm and store in R)
-            let mut norm_sq = T::zero();
-            for i in 0..m {
-                let val = q_data[i * min_mn + j];
-                norm_sq = norm_sq + val * val;
-            }
-            let norm = norm_sq.sqrt();
-            r_data[j * n + j] = norm;
-
-            // Normalize Q column
-            if norm > T::from(1e-10).unwrap_or(T::zero()) {
-                for i in 0..m {
-                    q_data[i * min_mn + j] = q_data[i * min_mn + j] / norm;
-                }
-            }
-        }
-
-        let q = Tensor::from_vec(q_data, vec![m, min_mn]);
-        let r = Tensor::from_vec(r_data, vec![min_mn, n]);
-
-        Ok((q, r))
-    }
-
-    /// Basic LU decomposition using Gaussian elimination
-    fn lu_basic(&self, m: usize, n: usize, min_mn: usize) -> Result<(Self, Self, Self), String> {
-        let matrix = self.as_array();
-
-        // Copy matrix data for in-place decomposition
-        let mut lu_data: Vec<T> = matrix.iter().cloned().collect();
-        let mut perm = vec![0usize; m];
-        for i in 0..m {
-            perm[i] = i;
-        }
-
-        // Gaussian elimination with partial pivoting
-        for k in 0..min_mn {
-            // Find pivot
-            let mut max_val = T::zero();
-            let mut pivot_row = k;
-
-            for i in k..m {
-                let abs_val = if lu_data[i * n + k] >= T::zero() {
-                    lu_data[i * n + k]
-                } else {
-                    T::zero() - lu_data[i * n + k]
-                };
-                if abs_val > max_val {
-                    max_val = abs_val;
-                    pivot_row = i;
-                }
-            }
-
-            // Swap rows if needed
-            if pivot_row != k {
-                for j in 0..n {
-                    let temp = lu_data[k * n + j];
-                    lu_data[k * n + j] = lu_data[pivot_row * n + j];
-                    lu_data[pivot_row * n + j] = temp;
-                }
-                perm.swap(k, pivot_row);
-            }
-
-            // Check for zero pivot
-            let pivot = lu_data[k * n + k];
-            if pivot.abs() < T::from(1e-10).unwrap_or(T::zero()) {
-                continue; // Skip if pivot is too small
-            }
-
-            // Elimination
-            for i in (k + 1)..m {
-                let factor = lu_data[i * n + k] / pivot;
-                lu_data[i * n + k] = factor; // Store L factor
-
-                for j in (k + 1)..n {
-                    lu_data[i * n + j] = lu_data[i * n + j] - factor * lu_data[k * n + j];
-                }
-            }
-        }
-
-        // Extract L and U matrices
-        let mut l_data = vec![T::zero(); m * min_mn];
-        let mut u_data = vec![T::zero(); min_mn * n];
-
-        for i in 0..m {
-            for j in 0..min_mn {
-                if i == j {
-                    l_data[i * min_mn + j] = T::one(); // Diagonal is 1
-                } else if i > j {
-                    l_data[i * min_mn + j] = lu_data[i * n + j]; // Below diagonal
-                }
-            }
-        }
-
-        for i in 0..min_mn {
-            for j in 0..n {
-                if i <= j {
-                    u_data[i * n + j] = lu_data[i * n + j]; // Upper triangular
-                }
-            }
-        }
-
-        // Create permutation matrix
-        let mut p_data = vec![T::zero(); m * m];
-        for i in 0..m {
-            p_data[i * m + perm[i]] = T::one();
-        }
-
-        let l = Tensor::from_vec(l_data, vec![m, min_mn]);
-        let u = Tensor::from_vec(u_data, vec![min_mn, n]);
-        let p = Tensor::from_vec(p_data, vec![m, m]);
-
-        Ok((l, u, p))
-    }
-
-    /// QR decomposition using ndarray-linalg
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn qr_with_linalg(&self, _m: usize, _n: usize, _min_mn: usize) -> Result<(Self, Self), String> {
-        use ndarray_linalg::QR;
-
-        let matrix = self.as_array().clone();
-
-        // Convert to 2D array for linalg operations
-        let matrix_2d = matrix
-            .into_dimensionality::<ndarray::Ix2>()
-            .map_err(|e| format!("Failed to convert to 2D array: {:?}", e))?;
-
-        // Convert to f64 for computation
-        let matrix_f64 = matrix_2d.mapv(|x| x.to_f64().unwrap_or(0.0));
-
-        let (q, r) = matrix_f64
-            .qr()
-            .map_err(|e| format!("QR decomposition failed: {:?}", e))?;
-
-        // Convert back to tensor type
-        let q_data: Vec<T> = q.iter().map(|&x| T::from(x).unwrap_or(T::zero())).collect();
-        let r_data: Vec<T> = r.iter().map(|&x| T::from(x).unwrap_or(T::zero())).collect();
-
-        let q_tensor = Tensor::from_vec(q_data, vec![q.nrows(), q.ncols()]);
-        let r_tensor = Tensor::from_vec(r_data, vec![r.nrows(), r.ncols()]);
-
-        Ok((q_tensor, r_tensor))
-    }
-
-    /// LU decomposition using ndarray-linalg
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn lu_with_linalg(
-        &self,
-        m: usize,
-        n: usize,
-        min_mn: usize,
-    ) -> Result<(Self, Self, Self), String> {
-        // ndarray-linalg doesn't provide LU decomposition directly
-        // Fall back to basic implementation
-        self.lu_basic(m, n, min_mn)
-    }
-
-    /// Sum along a specific axis.
-    /// 特定の軸に沿った合計
-    pub fn sum_axis(&self, axis: usize) -> Result<Self, String> {
-        if axis >= self.ndim() {
-            return Err(format!(
-                "Axis {} out of bounds for {}-dimensional tensor",
-                axis,
-                self.ndim()
-            ));
-        }
-
-        let sum_result = self.as_array().sum_axis(ndarray::Axis(axis));
-        Ok(Tensor::new(sum_result))
-    }
-
-    /// Mean along a specific axis.
-    /// 特定の軸に沿った平均
-    pub fn mean_axis(&self, axis: usize) -> Result<Self, String> {
-        if axis >= self.ndim() {
-            return Err(format!(
-                "Axis {} out of bounds for {}-dimensional tensor",
-                axis,
-                self.ndim()
-            ));
-        }
-
-        let mean_result = self.as_array().mean_axis(ndarray::Axis(axis)).unwrap();
-        Ok(Tensor::new(mean_result))
-    }
-
-    /// Apply a function element-wise.
-    ///
-    /// Apply exponential function element-wise.
-    /// 要素ごとに指数関数を適用
-    pub fn exp(&self) -> Self {
-        self.map(|x| x.exp())
-    }
-
-    /// Apply natural logarithm element-wise.
-    /// 要素ごとに自然対数を適用
-    pub fn ln(&self) -> Self {
-        self.map(|x| x.ln())
-    }
-
-    /// Apply sine function element-wise.
-    /// 要素ごとにサイン関数を適用
-    pub fn sin(&self) -> Self {
-        self.map(|x| x.sin())
-    }
-
-    /// Apply cosine function element-wise.
-    /// 要素ごとにコサイン関数を適用
-    pub fn cos(&self) -> Self {
-        self.map(|x| x.cos())
-    }
-
-    /// Apply tangent function element-wise.
-    /// 要素ごとにタンジェント関数を適用
-    pub fn tan(&self) -> Self {
-        self.map(|x| x.tan())
-    }
-
-    /// Apply square root element-wise.
-    /// 要素ごとに平方根を適用
-    pub fn sqrt(&self) -> Self {
-        self.map(|x| x.sqrt())
-    }
-
-    /// Apply absolute value element-wise.
-    /// 要素ごとに絶対値を適用
-    pub fn abs(&self) -> Self {
-        self.map(|x| x.abs())
-    }
-
-    /// Power operation element-wise.
-    /// 要素ごとのべき乗演算
-    pub fn pow(&self, exponent: T) -> Self {
-        self.map(|x| x.powf(exponent))
-    }
-}
-
-// Operator overloading for tensors
-impl<T: Float + 'static + ndarray::ScalarOperand + num_traits::FromPrimitive> ops::Add
-    for &Tensor<T>
-{
-    type Output = Tensor<T>;
-
-    fn add(self, rhs: Self) -> Self::Output {
-        self.add(rhs).expect("Addition failed")
-    }
-}
-
-impl<T: Float + 'static + ndarray::ScalarOperand + num_traits::FromPrimitive> ops::Sub
-    for &Tensor<T>
-{
-    type Output = Tensor<T>;
-
-    fn sub(self, rhs: Self) -> Self::Output {
-        self.sub(rhs).expect("Subtraction failed")
-    }
-}
-
-impl<T: Float + 'static + ndarray::ScalarOperand + num_traits::FromPrimitive> ops::Mul
-    for &Tensor<T>
-{
-    type Output = Tensor<T>;
-
-    fn mul(self, rhs: Self) -> Self::Output {
-        self.mul(rhs).expect("Multiplication failed")
-    }
-}
-
-impl<T: Float + 'static + ndarray::ScalarOperand + num_traits::FromPrimitive> ops::Div
-    for &Tensor<T>
-{
-    type Output = Tensor<T>;
-
-    fn div(self, rhs: Self) -> Self::Output {
-        self.div(rhs).expect("Division failed")
-    }
-}
-
-impl<
-        T: Float
-            + 'static
-            + ndarray::ScalarOperand
-            + num_traits::FromPrimitive
-            + std::ops::Neg<Output = T>,
-    > ops::Neg for &Tensor<T>
-{
-    type Output = Tensor<T>;
-
-    fn neg(self) -> Self::Output {
-        let result_data = self.as_array().mapv(|x| -x);
-        Tensor::new(result_data)
-    }
-}
-
-// Scalar operations
-impl<T: Float + 'static + ndarray::ScalarOperand + num_traits::FromPrimitive> ops::Add<T>
-    for &Tensor<T>
-{
-    type Output = Tensor<T>;
-
-    fn add(self, rhs: T) -> Self::Output {
-        self.add_scalar(rhs)
-    }
-}
-
-impl<T: Float + 'static + ndarray::ScalarOperand + num_traits::FromPrimitive> ops::Sub<T>
-    for &Tensor<T>
-{
-    type Output = Tensor<T>;
-
-    fn sub(self, rhs: T) -> Self::Output {
-        self.sub_scalar(rhs)
-    }
-}
-
-impl<T: Float + 'static + ndarray::ScalarOperand + num_traits::FromPrimitive> ops::Mul<T>
-    for &Tensor<T>
-{
-    type Output = Tensor<T>;
-
-    fn mul(self, rhs: T) -> Self::Output {
-        self.mul_scalar(rhs)
-    }
-}
-
-impl<T: Float + 'static + ndarray::ScalarOperand + num_traits::FromPrimitive> ops::Div<T>
-    for &Tensor<T>
-{
-    type Output = Tensor<T>;
-
-    fn div(self, rhs: T) -> Self::Output {
-        self.div_scalar(rhs)
-    }
-}
-
-// In-place operations
-impl<T: Float + 'static + ndarray::ScalarOperand + num_traits::FromPrimitive>
-    ops::AddAssign<&Tensor<T>> for Tensor<T>
-{
-    fn add_assign(&mut self, rhs: &Tensor<T>) {
-        *self = self.add(rhs).expect("AddAssign failed");
-    }
-}
-
-impl<T: Float + 'static + ndarray::ScalarOperand + num_traits::FromPrimitive>
-    ops::SubAssign<&Tensor<T>> for Tensor<T>
-{
-    fn sub_assign(&mut self, rhs: &Tensor<T>) {
-        *self = self.sub(rhs).expect("SubAssign failed");
-    }
-}
-
-impl<T: Float + 'static + ndarray::ScalarOperand + num_traits::FromPrimitive>
-    ops::MulAssign<&Tensor<T>> for Tensor<T>
-{
-    fn mul_assign(&mut self, rhs: &Tensor<T>) {
-        *self = self.mul(rhs).expect("MulAssign failed");
-    }
-}
-
-impl<T: Float + 'static + ndarray::ScalarOperand + num_traits::FromPrimitive>
-    ops::DivAssign<&Tensor<T>> for Tensor<T>
-{
-    fn div_assign(&mut self, rhs: &Tensor<T>) {
-        *self = self.div(rhs).expect("DivAssign failed");
-    }
-}
-
-// Value-based operations for direct tensor operations
-impl<T: Float + 'static + ndarray::ScalarOperand + num_traits::FromPrimitive> ops::Add
-    for Tensor<T>
-{
-    type Output = Tensor<T>;
-
-    fn add(self, rhs: Self) -> Self::Output {
-        (&self).add(&rhs).expect("Addition failed")
-    }
-}
-
-impl<T: Float + 'static + ndarray::ScalarOperand + num_traits::FromPrimitive> ops::Sub
-    for Tensor<T>
-{
-    type Output = Tensor<T>;
-
-    fn sub(self, rhs: Self) -> Self::Output {
-        (&self).sub(&rhs).expect("Subtraction failed")
-    }
-}
-
-impl<T: Float + 'static + ndarray::ScalarOperand + num_traits::FromPrimitive> ops::Mul
-    for Tensor<T>
-{
-    type Output = Tensor<T>;
-
-    fn mul(self, rhs: Self) -> Self::Output {
-        (&self).mul(&rhs).expect("Multiplication failed")
-    }
-}
-
-impl<T: Float + 'static + ndarray::ScalarOperand + num_traits::FromPrimitive> ops::Div
-    for Tensor<T>
-{
-    type Output = Tensor<T>;
-
-    fn div(self, rhs: Self) -> Self::Output {
-        (&self).div(&rhs).expect("Division failed")
-    }
-}
-
-// Scalar operations with tensor values
-impl<T: Float + 'static + ndarray::ScalarOperand + num_traits::FromPrimitive> ops::Add<T>
-    for Tensor<T>
-{
-    type Output = Tensor<T>;
-
-    fn add(self, rhs: T) -> Self::Output {
-        (&self).add_scalar(rhs)
-    }
-}
-
-impl<T: Float + 'static + ndarray::ScalarOperand + num_traits::FromPrimitive> ops::Sub<T>
-    for Tensor<T>
-{
-    type Output = Tensor<T>;
-
-    fn sub(self, rhs: T) -> Self::Output {
-        (&self).sub_scalar(rhs)
-    }
-}
-
-impl<T: Float + 'static + ndarray::ScalarOperand + num_traits::FromPrimitive> ops::Mul<T>
-    for Tensor<T>
-{
-    type Output = Tensor<T>;
-
-    fn mul(self, rhs: T) -> Self::Output {
-        (&self).mul_scalar(rhs)
-    }
-}
-
-impl<T: Float + 'static + ndarray::ScalarOperand + num_traits::FromPrimitive> ops::Div<T>
-    for Tensor<T>
-{
-    type Output = Tensor<T>;
-
-    fn div(self, rhs: T) -> Self::Output {
-        (&self).div_scalar(rhs)
-    }
-}
-
-// Mixed reference/value operations
-impl<T: Float + 'static + ndarray::ScalarOperand + num_traits::FromPrimitive> ops::Add<&Tensor<T>>
-    for Tensor<T>
-{
-    type Output = Tensor<T>;
-
-    fn add(self, rhs: &Tensor<T>) -> Self::Output {
-        (&self).add(rhs).expect("Addition failed")
-    }
-}
-
-impl<T: Float + 'static + ndarray::ScalarOperand + num_traits::FromPrimitive> ops::Add<Tensor<T>>
-    for &Tensor<T>
-{
-    type Output = Tensor<T>;
-
-    fn add(self, rhs: Tensor<T>) -> Self::Output {
-        self.add(&rhs).expect("Addition failed")
-    }
-}
-
-impl<T: Float + 'static + ndarray::ScalarOperand + num_traits::FromPrimitive> Tensor<T> {
-    // ===== Fourier Transform Operations =====
-    // torch.fft.* compatible API
-    // フーリエ変換演算（torch.fft.*互換API）
-
-    /// Fast Fourier Transform (FFT) - torch.fft.fft compatible
-    /// 高速フーリエ変換 - torch.fft.fft互換
-    ///
-    /// Returns (real_part, imaginary_part) as separate tensors
-    /// 実部と虚部を別々のテンソルとして返す
-    ///
-    /// # Arguments
-    /// * `n` - Length of the transformed axis. If None, use the input length.
-    /// * `dim` - Dimension along which to take the FFT. Default: -1 (last dimension)
-    /// * `norm` - Normalization mode: None, "forward", "backward", "ortho"
-    pub fn fft(
-        &self,
-        n: Option<usize>,
-        dim: Option<isize>,
-        norm: Option<&str>,
-    ) -> Result<(Self, Self), String>
-    where
-        T: Float + 'static + Default + Clone + std::fmt::Debug + num_traits::FromPrimitive,
-    {
-        let actual_dim = self.resolve_dim(dim.unwrap_or(-1))?;
-        let input_len = self.shape()[actual_dim];
-        let fft_len = n.unwrap_or(input_len);
-
-        // For 1D case on the last dimension
-        if self.shape().len() == 1 && actual_dim == 0 {
-            return self.fft_1d_basic(fft_len, norm, false);
-        }
-
-        Err("Multi-dimensional FFT not implemented in this version".to_string())
-    }
-
-    /// Inverse Fast Fourier Transform (IFFT) - torch.fft.ifft compatible
-    /// 逆高速フーリエ変換 - torch.fft.ifft互換
-    pub fn ifft(
-        &self,
-        real_part: &Self,
-        imag_part: &Self,
-        n: Option<usize>,
-        dim: Option<isize>,
-        norm: Option<&str>,
-    ) -> Result<(Self, Self), String>
-    where
-        T: Float + 'static + Default + Clone + std::fmt::Debug + num_traits::FromPrimitive,
-    {
-        let actual_dim = self.resolve_dim(dim.unwrap_or(-1))?;
-        let input_len = self.shape()[actual_dim];
-        let fft_len = n.unwrap_or(input_len);
-
-        // For 1D case on the last dimension
-        if self.shape().len() == 1 && actual_dim == 0 {
-            return self.ifft_1d_basic(real_part, imag_part, fft_len, norm);
-        }
-
-        Err("Multi-dimensional IFFT not implemented in this version".to_string())
-    }
-
-    /// Real FFT - torch.fft.rfft compatible
-    /// 実数FFT - torch.fft.rfft互換
-    pub fn rfft(
-        &self,
-        n: Option<usize>,
-        dim: Option<isize>,
-        norm: Option<&str>,
-    ) -> Result<(Self, Self), String>
-    where
-        T: Float + 'static + Default + Clone + std::fmt::Debug + num_traits::FromPrimitive,
-    {
-        // RFFT returns only the first N/2+1 frequencies due to Hermitian symmetry
-        let (real_full, imag_full) = self.fft(n, dim, norm)?;
-        let full_len = real_full.shape()[0];
-        let rfft_len = full_len / 2 + 1;
-
-        // Extract first half + 1
-        let real_data = real_full.data.as_slice().unwrap();
-        let imag_data = imag_full.data.as_slice().unwrap();
-
-        let real_rfft: Vec<T> = real_data[0..rfft_len].to_vec();
-        let imag_rfft: Vec<T> = imag_data[0..rfft_len].to_vec();
-
-        Ok((
-            Tensor::from_vec(real_rfft, vec![rfft_len]),
-            Tensor::from_vec(imag_rfft, vec![rfft_len]),
-        ))
-    }
-
-    /// FFT shift - torch.fft.fftshift compatible
-    /// FFTシフト - torch.fft.fftshift互換
-    pub fn fftshift(&self, dim: Option<&[isize]>) -> Result<Self, String> {
-        let actual_dim = if dim.is_some() {
-            self.resolve_dim(dim.unwrap()[0])?
-        } else {
-            self.shape().len() - 1 // Default to last dimension
-        };
-
-        let size = self.shape()[actual_dim];
-        let mid = size.div_ceil(2);
-
-        let input_data = self.data.as_slice().unwrap();
-        let mut new_data = Vec::new();
-
-        // For 1D case: shift second half to front, first half to back
-        if self.shape().len() == 1 {
-            new_data.extend_from_slice(&input_data[mid..]);
-            new_data.extend_from_slice(&input_data[..mid]);
-        } else {
-            return Err("Multi-dimensional fftshift not implemented in this version".to_string());
-        }
-
-        Ok(Tensor::from_vec(new_data, self.shape().to_vec()))
-    }
-
-    /// Inverse FFT shift - torch.fft.ifftshift compatible
-    /// 逆FFTシフト - torch.fft.ifftshift互換
-    pub fn ifftshift(&self, dim: Option<&[isize]>) -> Result<Self, String> {
-        let actual_dim = if dim.is_some() {
-            self.resolve_dim(dim.unwrap()[0])?
-        } else {
-            self.shape().len() - 1 // Default to last dimension
-        };
-
-        let size = self.shape()[actual_dim];
-        let mid = size / 2;
-
-        let input_data = self.data.as_slice().unwrap();
-        let mut new_data = Vec::new();
-
-        // For 1D case: shift second half to front, first half to back (different split)
-        if self.shape().len() == 1 {
-            new_data.extend_from_slice(&input_data[mid..]);
-            new_data.extend_from_slice(&input_data[..mid]);
-        } else {
-            return Err("Multi-dimensional ifftshift not implemented in this version".to_string());
-        }
-
-        Ok(Tensor::from_vec(new_data, self.shape().to_vec()))
-    }
-
-    // ===== Helper Methods for FFT =====
-
-    /// Basic 1D FFT implementation
-    /// 基本的な1D FFT実装
-    fn fft_1d_basic(
-        &self,
-        n: usize,
-        norm: Option<&str>,
-        inverse: bool,
-    ) -> Result<(Self, Self), String>
-    where
-        T: Float + 'static + Default + Clone + std::fmt::Debug + num_traits::FromPrimitive,
-    {
-        let input_data = self.data.as_slice().unwrap();
-        let input_len = input_data.len();
-
-        // Pad or truncate to desired length
-        let real_input: Vec<T> = if n != input_len {
-            if n > input_len {
-                // Zero-pad
-                let mut padded = input_data.to_vec();
-                padded.resize(n, T::zero());
-                padded
-            } else {
-                // Truncate
-                input_data[0..n].to_vec()
-            }
-        } else {
-            input_data.to_vec()
-        };
-
-        // Convert to complex representation
-        let mut complex_data: Vec<Complex<T>> = real_input
-            .iter()
-            .map(|&x| Complex::new(x, T::zero()))
-            .collect();
-
-        // Perform DFT (simple implementation)
-        let result = if n.is_power_of_two() {
-            self.cooley_tukey_1d(&mut complex_data, inverse)?
-        } else {
-            self.dft_1d(&complex_data, inverse)?
-        };
-
-        // Apply normalization
-        let normalized_result = self.apply_normalization(result, n, norm, inverse)?;
-
-        // Extract real and imaginary parts
-        let real_part: Vec<T> = normalized_result.iter().map(|c| c.re).collect();
-        let imag_part: Vec<T> = normalized_result.iter().map(|c| c.im).collect();
-
-        Ok((
-            Tensor::from_vec(real_part, vec![n]),
-            Tensor::from_vec(imag_part, vec![n]),
-        ))
-    }
-
-    /// Basic 1D IFFT implementation
-    /// 基本的な1D IFFT実装
-    fn ifft_1d_basic(
-        &self,
-        real_part: &Self,
-        imag_part: &Self,
-        n: usize,
-        norm: Option<&str>,
-    ) -> Result<(Self, Self), String>
-    where
-        T: Float + 'static + Default + Clone + std::fmt::Debug + num_traits::FromPrimitive,
-    {
-        // Reconstruct complex data
-        let real_data = real_part.data.as_slice().unwrap();
-        let imag_data = imag_part.data.as_slice().unwrap();
-
-        let mut complex_data: Vec<Complex<T>> = real_data
-            .iter()
-            .zip(imag_data.iter())
-            .map(|(&r, &i)| Complex::new(r, i))
-            .collect();
-
-        // Perform inverse DFT
-        let result = if n.is_power_of_two() {
-            self.cooley_tukey_1d(&mut complex_data, true)?
-        } else {
-            self.dft_1d(&complex_data, true)?
-        };
-
-        // Apply normalization
-        let normalized_result = self.apply_normalization(result, n, norm, true)?;
-
-        // Extract real and imaginary parts
-        let real_part: Vec<T> = normalized_result.iter().map(|c| c.re).collect();
-        let imag_part: Vec<T> = normalized_result.iter().map(|c| c.im).collect();
-
-        Ok((
-            Tensor::from_vec(real_part, vec![n]),
-            Tensor::from_vec(imag_part, vec![n]),
-        ))
-    }
-
-    /// Simple Cooley-Tukey FFT for power-of-2 sizes
-    /// 2の累乗サイズ用のシンプルなCooley-Tukey FFT
-    fn cooley_tukey_1d(
-        &self,
-        data: &mut [Complex<T>],
-        inverse: bool,
-    ) -> Result<Vec<Complex<T>>, String>
-    where
-        T: Float + 'static + Default + Clone + std::fmt::Debug + num_traits::FromPrimitive,
-    {
-        let n = data.len();
-        if n <= 1 {
-            return Ok(data.to_vec());
-        }
-
-        // Bit-reversal permutation
-        let mut j = 0;
-        for i in 1..n {
-            let mut bit = n >> 1;
-            while j & bit != 0 {
-                j ^= bit;
-                bit >>= 1;
-            }
-            j ^= bit;
-            if i < j {
-                data.swap(i, j);
-            }
-        }
-
-        // Cooley-Tukey FFT
-        let mut length = 2;
-        while length <= n {
-            let half_len = length / 2;
-            let angle = if inverse {
-                T::from(2.0).unwrap()
-            } else {
-                T::from(-2.0).unwrap()
-            } * T::from(std::f64::consts::PI).unwrap()
-                / T::from(length).unwrap();
-
-            let w_len = Complex::new(angle.cos(), angle.sin());
-
-            for i in (0..n).step_by(length) {
-                let mut w = Complex::new(T::one(), T::zero());
-
-                for j in 0..half_len {
-                    let u = data[i + j];
-                    let v = data[i + j + half_len] * w;
-                    data[i + j] = u + v;
-                    data[i + j + half_len] = u - v;
-                    w = w * w_len;
-                }
-            }
-
-            length <<= 1;
-        }
-
-        Ok(data.to_vec())
-    }
-
-    /// Simple DFT implementation for arbitrary sizes
-    /// 任意のサイズ用のシンプルなDFT実装
-    fn dft_1d(&self, data: &[Complex<T>], inverse: bool) -> Result<Vec<Complex<T>>, String>
-    where
-        T: Float + 'static + Default + Clone + std::fmt::Debug + num_traits::FromPrimitive,
-    {
-        let n = data.len();
-        let mut result = vec![Complex::new(T::zero(), T::zero()); n];
-
-        for k in 0..n {
-            let mut sum = Complex::new(T::zero(), T::zero());
-
-            for j in 0..n {
-                let angle = if inverse {
-                    T::from(2.0).unwrap()
-                } else {
-                    T::from(-2.0).unwrap()
-                } * T::from(std::f64::consts::PI).unwrap()
-                    * T::from(k).unwrap()
-                    * T::from(j).unwrap()
-                    / T::from(n).unwrap();
-
-                let twiddle = Complex::new(angle.cos(), angle.sin());
-                sum = sum + data[j] * twiddle;
-            }
-
-            result[k] = sum;
-        }
-
-        Ok(result)
-    }
-
-    /// Apply FFT normalization
-    /// FFT正規化を適用
-    fn apply_normalization(
-        &self,
-        mut data: Vec<Complex<T>>,
-        n: usize,
-        norm: Option<&str>,
-        inverse: bool,
-    ) -> Result<Vec<Complex<T>>, String>
-    where
-        T: Float + 'static + Default + Clone + std::fmt::Debug + num_traits::FromPrimitive,
-    {
-        match norm {
-            Some("forward") if !inverse => {
-                let scale = T::from(n).unwrap_or(T::one());
-                for c in &mut data {
-                    *c = *c / Complex::new(scale, T::zero());
-                }
-            }
-            Some("backward") if inverse => {
-                let scale = T::from(n).unwrap_or(T::one());
-                for c in &mut data {
-                    *c = *c / Complex::new(scale, T::zero());
-                }
-            }
-            Some("ortho") => {
-                let scale = T::from(n).unwrap_or(T::one()).sqrt();
-                for c in &mut data {
-                    *c = *c / Complex::new(scale, T::zero());
-                }
-            }
-            _ if inverse => {
-                // Default normalization for inverse transform
-                let scale = T::from(n).unwrap_or(T::one());
-                for c in &mut data {
-                    *c = *c / Complex::new(scale, T::zero());
-                }
-            }
-            _ => {} // No normalization
-        }
-
-        Ok(data)
-    }
-
-    /// Resolve negative dimension indices
-    /// 負の次元インデックスを解決
-    fn resolve_dim(&self, dim: isize) -> Result<usize, String> {
-        let ndim = self.shape().len() as isize;
-
-        let resolved = if dim < 0 { ndim + dim } else { dim };
-
-        if resolved < 0 || resolved >= ndim {
-            Err(format!(
-                "Dimension {} out of range for tensor with {} dimensions",
-                dim, ndim
-            ))
-        } else {
-            Ok(resolved as usize)
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::tensor::Tensor;
-    use approx::assert_abs_diff_eq;
-
-    #[test]
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn test_svd_square_matrix() {
-        // Test SVD on a simple 2x2 matrix
-        let matrix = Tensor::from_vec(vec![3.0f32, 1.0, 1.0, 3.0], vec![2, 2]);
-
-        let result = matrix.svd(true);
-        assert!(result.is_ok(), "SVD should succeed");
-
-        let (u, s, v) = result.unwrap();
-
-        // Check dimensions
-        assert_eq!(u.shape(), vec![2, 2]);
-        assert_eq!(s.shape(), vec![2]);
-        assert_eq!(v.shape(), vec![2, 2]);
-
-        // Check singular values are non-negative and sorted in descending order
-        let s_data = s.data.as_slice().unwrap();
-        assert!(s_data[0] >= 0.0);
-        assert!(s_data[1] >= 0.0);
-        assert!(s_data[0] >= s_data[1]);
-
-        // Note: Full reconstruction test A = U * S * V^T would require diag() method
-        // For now, we verify the basic properties of the SVD decomposition
-    }
-
-    #[test]
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn test_svd_rectangular_matrix() {
-        // Test SVD on a 3x2 rectangular matrix
-        let matrix = Tensor::from_vec(vec![1.0f32, 2.0, 3.0, 4.0, 5.0, 6.0], vec![3, 2]);
-
-        let result = matrix.svd(true);
-        assert!(result.is_ok(), "SVD should succeed for rectangular matrix");
-
-        let (u, s, v) = result.unwrap();
-
-        // For m x n matrix where m > n, expect:
-        // U: m x min(m,n) = 3 x 2
-        // S: min(m,n) = 2
-        // V: n x min(m,n) = 2 x 2
-        assert_eq!(u.shape(), vec![3, 2]);
-        assert_eq!(s.shape(), vec![2]);
-        assert_eq!(v.shape(), vec![2, 2]);
-
-        // Check singular values are sorted in descending order
-        let s_data = s.data.as_slice().unwrap();
-        assert!(s_data[0] >= s_data[1]);
-        assert!(s_data[0] >= 0.0);
-        assert!(s_data[1] >= 0.0);
-    }
-
-    #[test]
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn test_svd_orthogonality() {
-        // Test that U and V are approximately orthogonal matrices
-        // Note: Basic SVD implementation may not provide perfect orthogonality
-        let matrix = Tensor::from_vec(vec![2.0f32, 1.0, 1.0, 2.0], vec![2, 2]);
-
-        let result = matrix.svd(true);
-        assert!(result.is_ok());
-
-        let (u, _s, v) = result.unwrap();
-
-        // For basic implementation, we just verify the shapes and that
-        // the matrices are reasonable (not testing strict orthogonality)
-        assert_eq!(u.shape(), vec![2, 2]);
-        assert_eq!(v.shape(), vec![2, 2]);
-
-        // Verify that U and V contain finite values
-        let u_data = u.data.as_slice().unwrap();
-        let v_data = v.data.as_slice().unwrap();
-
-        for &val in u_data {
-            assert!(val.is_finite());
-        }
-        for &val in v_data {
-            assert!(val.is_finite());
-        }
-    }
-
-    #[test]
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn test_svd_rank_deficient() {
-        // Test SVD on a rank-deficient matrix (rank 1)
-        let matrix = Tensor::from_vec(vec![1.0f32, 2.0, 2.0, 4.0], vec![2, 2]);
-
-        let result = matrix.svd(true);
-        assert!(result.is_ok());
-
-        let (_u, s, _v) = result.unwrap();
-        let s_data = s.data.as_slice().unwrap();
-
-        // One singular value should be significantly larger than the other
-        // (indicating rank deficiency)
-        // Note: basic implementation may not detect rank deficiency perfectly
-        assert!(s_data[0] >= s_data[1]); // Should be sorted in descending order
-        assert!(s_data[0] > 0.0); // At least one non-zero singular value
-    }
-
-    #[test]
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn test_svd_identity_matrix() {
-        // Test SVD on identity matrix
-        let matrix = Tensor::from_vec(vec![1.0f32, 0.0, 0.0, 1.0], vec![2, 2]);
-
-        let result = matrix.svd(true);
-        assert!(result.is_ok());
-
-        let (_u, s, _v) = result.unwrap();
-        let s_data = s.data.as_slice().unwrap();
-
-        // Singular values of identity should be 1.0
-        assert_abs_diff_eq!(s_data[0], 1.0, epsilon = 1e-5);
-        assert_abs_diff_eq!(s_data[1], 1.0, epsilon = 1e-5);
-    }
-
-    #[test]
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn test_svd_zero_matrix() {
-        // Test SVD on zero matrix
-        let matrix = Tensor::from_vec(vec![0.0f32, 0.0, 0.0, 0.0], vec![2, 2]);
-
-        let result = matrix.svd(true);
-        assert!(result.is_ok());
-
-        let (_u, s, _v) = result.unwrap();
-        let s_data = s.data.as_slice().unwrap();
-
-        // All singular values should be zero
-        assert_abs_diff_eq!(s_data[0], 0.0, epsilon = 1e-6);
-        assert_abs_diff_eq!(s_data[1], 0.0, epsilon = 1e-6);
-    }
-
-    #[test]
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn test_svd_some_false() {
-        // Test SVD with some=false (reduced SVD)
-        let matrix = Tensor::from_vec(vec![1.0f32, 2.0, 3.0, 4.0, 5.0, 6.0], vec![3, 2]);
-
-        let result = matrix.svd(false);
-        assert!(result.is_ok());
-
-        let (u, s, v) = result.unwrap();
-
-        // Basic implementation always returns reduced form
-        // For 3x2 matrix: U should be 3x2, S should be 2, V should be 2x2
-        assert_eq!(u.shape(), vec![3, 2]);
-        assert_eq!(s.shape(), vec![2]);
-        assert_eq!(v.shape(), vec![2, 2]);
-    }
-
-    #[test]
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn test_svd_tall_thin_matrix() {
-        // Test SVD on a tall, thin matrix (4x2)
-        let matrix = Tensor::from_vec(vec![1.0f32, 0.0, 0.0, 1.0, 1.0, 0.0, 0.0, 1.0], vec![4, 2]);
-
-        let result = matrix.svd(true);
-        assert!(result.is_ok());
-
-        let (u, s, v) = result.unwrap();
-
-        // Check dimensions for tall thin matrix
-        assert_eq!(u.shape(), vec![4, 2]);
-        assert_eq!(s.shape(), vec![2]);
-        assert_eq!(v.shape(), vec![2, 2]);
-
-        // Singular values should be positive and sorted
-        let s_data = s.data.as_slice().unwrap();
-        assert!(s_data[0] >= s_data[1]);
-        assert!(s_data[0] >= 0.0);
-        assert!(s_data[1] >= 0.0);
-    }
-
-    #[test]
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn test_svd_wide_matrix() {
-        // Test SVD on a wide matrix (2x4)
-        let matrix = Tensor::from_vec(vec![1.0f32, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 1.0], vec![2, 4]);
-
-        let result = matrix.svd(true);
-        assert!(result.is_ok());
-
-        let (u, s, v) = result.unwrap();
-
-        // Check dimensions for wide matrix
-        assert_eq!(u.shape(), vec![2, 2]);
-        assert_eq!(s.shape(), vec![2]);
-        assert_eq!(v.shape(), vec![4, 2]);
-
-        // Singular values should be positive and sorted
-        let s_data = s.data.as_slice().unwrap();
-        assert!(s_data[0] >= s_data[1]);
-        assert!(s_data[0] >= 0.0);
-        assert!(s_data[1] >= 0.0);
-    }
-
-    #[test]
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn test_eig_general_matrix() {
-        // Test general eigenvalue decomposition
-        let matrix = Tensor::from_vec(vec![2.0f32, 1.0, 1.0, 2.0], vec![2, 2]);
-
-        let result = matrix.eig(false);
-        assert!(
-            result.is_ok(),
-            "General eigenvalue decomposition should succeed"
-        );
-
-        let (eigenvals, eigenvecs) = result.unwrap();
-
-        // Check eigenvalues shape: [n, 2] for complex eigenvalues
-        assert_eq!(eigenvals.shape(), vec![2, 2]);
-        assert!(
-            eigenvecs.is_none(),
-            "Should not return eigenvectors when requested false"
-        );
-
-        // Check that eigenvalues are finite
-        let eigenvals_data = eigenvals.data.as_slice().unwrap();
-        for &val in eigenvals_data {
-            assert!(val.is_finite(), "Eigenvalues should be finite");
-        }
-    }
-
-    #[test]
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn test_eig_with_eigenvectors() {
-        // Test general eigenvalue decomposition with eigenvectors
-        let matrix = Tensor::from_vec(vec![3.0f32, 0.0, 0.0, 2.0], vec![2, 2]);
-
-        let result = matrix.eig(true);
-        assert!(
-            result.is_ok(),
-            "Eigenvalue decomposition with vectors should succeed"
-        );
-
-        let (eigenvals, eigenvecs) = result.unwrap();
-
-        // Check dimensions
-        assert_eq!(eigenvals.shape(), vec![2, 2]);
-        assert!(
-            eigenvecs.is_some(),
-            "Should return eigenvectors when requested"
-        );
-
-        if let Some(vecs) = eigenvecs {
-            assert_eq!(vecs.shape(), vec![2, 2]);
-
-            // Verify eigenvectors contain finite values
-            let vecs_data = vecs.data.as_slice().unwrap();
-            for &val in vecs_data {
-                assert!(val.is_finite(), "Eigenvector elements should be finite");
-            }
-        }
-    }
-
-    #[test]
-    fn test_symeig_symmetric_matrix() {
-        // Test symmetric eigenvalue decomposition
-        let matrix = Tensor::from_vec(vec![4.0f32, 1.0, 1.0, 3.0], vec![2, 2]);
-
-        let result = matrix.symeig(false, false);
-        assert!(
-            result.is_ok(),
-            "Symmetric eigenvalue decomposition should succeed"
-        );
-
-        let (eigenvals, eigenvecs) = result.unwrap();
-
-        // Check eigenvalues shape: [n] for real eigenvalues
-        assert_eq!(eigenvals.shape(), vec![2]);
-        assert!(
-            eigenvecs.is_none(),
-            "Should not return eigenvectors when requested false"
-        );
-
-        // Check eigenvalues are sorted (ascending by default)
-        let eigenvals_data = eigenvals.data.as_slice().unwrap();
-        assert!(
-            eigenvals_data[0] <= eigenvals_data[1],
-            "Eigenvalues should be sorted ascending"
-        );
-
-        // All eigenvalues should be finite
-        for &val in eigenvals_data {
-            assert!(val.is_finite(), "Eigenvalues should be finite");
-        }
-    }
-
-    #[test]
-    fn test_symeig_with_eigenvectors() {
-        // Test symmetric eigenvalue decomposition with eigenvectors
-        let matrix = Tensor::from_vec(vec![5.0f32, 0.0, 0.0, 3.0], vec![2, 2]);
-
-        let result = matrix.symeig(true, false);
-        assert!(
-            result.is_ok(),
-            "Symmetric eigenvalue decomposition with vectors should succeed"
-        );
-
-        let (eigenvals, eigenvecs) = result.unwrap();
-
-        // Check dimensions
-        assert_eq!(eigenvals.shape(), vec![2]);
-        assert!(
-            eigenvecs.is_some(),
-            "Should return eigenvectors when requested"
-        );
-
-        if let Some(vecs) = eigenvecs {
-            assert_eq!(vecs.shape(), vec![2, 2]);
-
-            // Verify eigenvectors are finite
-            let vecs_data = vecs.data.as_slice().unwrap();
-            for &val in vecs_data {
-                assert!(val.is_finite(), "Eigenvector elements should be finite");
-            }
-        }
-
-        // For diagonal matrix, eigenvalues should be the diagonal elements (sorted)
-        let eigenvals_data = eigenvals.data.as_slice().unwrap();
-        let mut expected = [3.0, 5.0]; // diagonal elements
-        expected.sort_by(|a, b| a.partial_cmp(b).unwrap());
-
-        assert_abs_diff_eq!(eigenvals_data[0], expected[0], epsilon = 1e-4);
-        assert_abs_diff_eq!(eigenvals_data[1], expected[1], epsilon = 1e-4);
-    }
-
-    #[test]
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn test_eig_error_cases() {
-        // Test with non-square matrix
-        let matrix = Tensor::from_vec(vec![1.0f32, 2.0, 3.0, 4.0, 5.0, 6.0], vec![2, 3]);
-
-        let result = matrix.eig(false);
-        assert!(result.is_err(), "Should fail for non-square matrix");
-
-        // Test with 1D tensor
-        let vector = Tensor::from_vec(vec![1.0f32, 2.0, 3.0], vec![3]);
-        let result = vector.eig(false);
-        assert!(result.is_err(), "Should fail for 1D tensor");
-    }
-
-    #[test]
-    fn test_symeig_error_cases() {
-        // Test with non-square matrix
-        let matrix = Tensor::from_vec(vec![1.0f32, 2.0, 3.0, 4.0, 5.0, 6.0], vec![2, 3]);
-
-        let result = matrix.symeig(false, false);
-        assert!(result.is_err(), "Should fail for non-square matrix");
-
-        // Test with 1D tensor
-        let vector = Tensor::from_vec(vec![1.0f32, 2.0, 3.0], vec![3]);
-        let result = vector.symeig(false, false);
-        assert!(result.is_err(), "Should fail for 1D tensor");
-    }
-
-    #[test]
-    fn test_symeig_identity_matrix() {
-        // Test symmetric eigenvalues for identity matrix
-        let identity = Tensor::from_vec(vec![1.0f32, 0.0, 0.0, 1.0], vec![2, 2]);
-
-        let result = identity.symeig(false, false);
-        assert!(result.is_ok(), "Identity matrix eigenvalues should succeed");
-
-        let (eigenvals, _) = result.unwrap();
-        let eigenvals_data = eigenvals.data.as_slice().unwrap();
-
-        // All eigenvalues of identity matrix should be 1
-        for &val in eigenvals_data {
-            assert_abs_diff_eq!(val, 1.0, epsilon = 1e-4);
-        }
-    }
-
-    #[test]
-    fn test_symeig_zero_matrix() {
-        // Test symmetric eigenvalues for zero matrix
-        let zero_matrix = Tensor::from_vec(vec![0.0f32, 0.0, 0.0, 0.0], vec![2, 2]);
-
-        let result = zero_matrix.symeig(false, false);
-        assert!(result.is_ok(), "Zero matrix eigenvalues should succeed");
-
-        let (eigenvals, _) = result.unwrap();
-        let eigenvals_data = eigenvals.data.as_slice().unwrap();
-
-        // All eigenvalues of zero matrix should be 0
-        for &val in eigenvals_data {
-            assert_abs_diff_eq!(val, 0.0, epsilon = 1e-6);
-        }
-    }
-
-    #[test]
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn test_qr_decomposition() {
-        // Test QR decomposition on a simple matrix
-        let matrix = Tensor::from_vec(vec![1.0f32, 2.0, 3.0, 4.0], vec![2, 2]);
-
-        let result = matrix.qr();
-        assert!(result.is_ok(), "QR decomposition should succeed");
-
-        let (q, r) = result.unwrap();
-
-        // Check dimensions
-        assert_eq!(q.shape(), vec![2, 2]);
-        assert_eq!(r.shape(), vec![2, 2]);
-
-        // Check that Q and R contain finite values
-        let q_data = q.data.as_slice().unwrap();
-        let r_data = r.data.as_slice().unwrap();
-
-        for &val in q_data {
-            assert!(val.is_finite(), "Q matrix elements should be finite");
-        }
-        for &val in r_data {
-            assert!(val.is_finite(), "R matrix elements should be finite");
-        }
-
-        // Check that R is upper triangular (lower triangle should be close to zero)
-        // R[1,0] should be approximately zero
-        assert!(r_data[2].abs() < 1e-4, "R should be upper triangular");
-    }
-
-    #[test]
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn test_qr_rectangular_matrix() {
-        // Test QR decomposition on rectangular matrix
-        let matrix = Tensor::from_vec(vec![1.0f32, 2.0, 3.0, 4.0, 5.0, 6.0], vec![3, 2]);
-
-        let result = matrix.qr();
-        assert!(
-            result.is_ok(),
-            "QR decomposition should succeed for rectangular matrix"
-        );
-
-        let (q, r) = result.unwrap();
-
-        // Check dimensions for 3x2 matrix
-        assert_eq!(q.shape(), vec![3, 2]);
-        assert_eq!(r.shape(), vec![2, 2]);
-
-        // Verify all values are finite
-        let q_data = q.data.as_slice().unwrap();
-        let r_data = r.data.as_slice().unwrap();
-
-        for &val in q_data.iter().chain(r_data.iter()) {
-            assert!(val.is_finite(), "QR decomposition results should be finite");
-        }
-    }
-
-    #[test]
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn test_lu_decomposition() {
-        // Test LU decomposition on a simple matrix
-        let matrix = Tensor::from_vec(vec![2.0f32, 1.0, 1.0, 3.0], vec![2, 2]);
-
-        let result = matrix.lu();
-        assert!(result.is_ok(), "LU decomposition should succeed");
-
-        let (l, u, p) = result.unwrap();
-
-        // Check dimensions
-        assert_eq!(l.shape(), vec![2, 2]);
-        assert_eq!(u.shape(), vec![2, 2]);
-        assert_eq!(p.shape(), vec![2, 2]);
-
-        // Check that L, U, P contain finite values
-        let l_data = l.data.as_slice().unwrap();
-        let u_data = u.data.as_slice().unwrap();
-        let p_data = p.data.as_slice().unwrap();
-
-        for &val in l_data.iter().chain(u_data.iter()).chain(p_data.iter()) {
-            assert!(val.is_finite(), "LU decomposition results should be finite");
-        }
-
-        // Check that L has unit diagonal
-        assert_abs_diff_eq!(l_data[0], 1.0, epsilon = 1e-5); // L[0,0]
-        assert_abs_diff_eq!(l_data[3], 1.0, epsilon = 1e-5); // L[1,1]
-
-        // Check that L is lower triangular (upper triangle should be zero)
-        assert_abs_diff_eq!(l_data[1], 0.0, epsilon = 1e-5); // L[0,1]
-
-        // Check that U is upper triangular (lower triangle should be zero)
-        assert_abs_diff_eq!(u_data[2], 0.0, epsilon = 1e-5); // U[1,0]
-    }
-
-    #[test]
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn test_lu_rectangular_matrix() {
-        // Test LU decomposition on rectangular matrix
-        let matrix = Tensor::from_vec(vec![1.0f32, 2.0, 3.0, 4.0, 5.0, 6.0], vec![3, 2]);
-
-        let result = matrix.lu();
-        assert!(
-            result.is_ok(),
-            "LU decomposition should succeed for rectangular matrix"
-        );
-
-        let (l, u, p) = result.unwrap();
-
-        // Check dimensions for 3x2 matrix
-        assert_eq!(l.shape(), vec![3, 2]);
-        assert_eq!(u.shape(), vec![2, 2]);
-        assert_eq!(p.shape(), vec![3, 3]);
-
-        // Verify all values are finite
-        let l_data = l.data.as_slice().unwrap();
-        let u_data = u.data.as_slice().unwrap();
-        let p_data = p.data.as_slice().unwrap();
-
-        for &val in l_data.iter().chain(u_data.iter()).chain(p_data.iter()) {
-            assert!(val.is_finite(), "LU decomposition results should be finite");
-        }
-    }
-
-    #[test]
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn test_qr_identity_matrix() {
-        // Test QR decomposition on identity matrix
-        let identity = Tensor::from_vec(vec![1.0f32, 0.0, 0.0, 1.0], vec![2, 2]);
-
-        let result = identity.qr();
-        assert!(
-            result.is_ok(),
-            "QR decomposition should succeed for identity matrix"
-        );
-
-        let (q, r) = result.unwrap();
-
-        // For identity matrix, Q should be close to identity and R should be close to identity
-        let q_data = q.data.as_slice().unwrap();
-        let r_data = r.data.as_slice().unwrap();
-
-        // Check Q is approximately identity
-        assert_abs_diff_eq!(q_data[0], 1.0, epsilon = 1e-4); // Q[0,0]
-        assert_abs_diff_eq!(q_data[1], 0.0, epsilon = 1e-4); // Q[0,1]
-        assert_abs_diff_eq!(q_data[2], 0.0, epsilon = 1e-4); // Q[1,0]
-        assert_abs_diff_eq!(q_data[3], 1.0, epsilon = 1e-4); // Q[1,1]
-
-        // Check R is approximately identity
-        assert_abs_diff_eq!(r_data[0], 1.0, epsilon = 1e-4); // R[0,0]
-        assert_abs_diff_eq!(r_data[1], 0.0, epsilon = 1e-4); // R[0,1]
-        assert_abs_diff_eq!(r_data[2], 0.0, epsilon = 1e-4); // R[1,0]
-        assert_abs_diff_eq!(r_data[3], 1.0, epsilon = 1e-4); // R[1,1]
-    }
-
-    #[test]
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn test_lu_identity_matrix() {
-        // Test LU decomposition on identity matrix
-        let identity = Tensor::from_vec(vec![1.0f32, 0.0, 0.0, 1.0], vec![2, 2]);
-
-        let result = identity.lu();
-        assert!(
-            result.is_ok(),
-            "LU decomposition should succeed for identity matrix"
-        );
-
-        let (l, u, p) = result.unwrap();
-
-        // For identity matrix, L should be identity, U should be identity, P should be identity
-        let l_data = l.data.as_slice().unwrap();
-        let u_data = u.data.as_slice().unwrap();
-        let p_data = p.data.as_slice().unwrap();
-
-        // Check L is approximately identity
-        assert_abs_diff_eq!(l_data[0], 1.0, epsilon = 1e-4); // L[0,0]
-        assert_abs_diff_eq!(l_data[1], 0.0, epsilon = 1e-4); // L[0,1]
-        assert_abs_diff_eq!(l_data[2], 0.0, epsilon = 1e-4); // L[1,0]
-        assert_abs_diff_eq!(l_data[3], 1.0, epsilon = 1e-4); // L[1,1]
-
-        // Check U is approximately identity
-        assert_abs_diff_eq!(u_data[0], 1.0, epsilon = 1e-4); // U[0,0]
-        assert_abs_diff_eq!(u_data[1], 0.0, epsilon = 1e-4); // U[0,1]
-        assert_abs_diff_eq!(u_data[2], 0.0, epsilon = 1e-4); // U[1,0]
-        assert_abs_diff_eq!(u_data[3], 1.0, epsilon = 1e-4); // U[1,1]
-
-        // Check P is approximately identity
-        assert_abs_diff_eq!(p_data[0], 1.0, epsilon = 1e-4); // P[0,0]
-        assert_abs_diff_eq!(p_data[1], 0.0, epsilon = 1e-4); // P[0,1]
-        assert_abs_diff_eq!(p_data[2], 0.0, epsilon = 1e-4); // P[1,0]
-        assert_abs_diff_eq!(p_data[3], 1.0, epsilon = 1e-4); // P[1,1]
-    }
-
-    #[test]
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn test_qr_error_cases() {
-        // Test with 1D tensor
-        let vector = Tensor::from_vec(vec![1.0f32, 2.0, 3.0], vec![3]);
-        let result = vector.qr();
-        assert!(result.is_err(), "Should fail for 1D tensor");
-    }
-
-    #[test]
-    #[cfg(all(
-        any(feature = "linalg", feature = "linalg-system"),
-        not(target_arch = "wasm32")
-    ))]
-    fn test_lu_error_cases() {
-        // Test with 1D tensor
-        let vector = Tensor::from_vec(vec![1.0f32, 2.0, 3.0], vec![3]);
-        let result = vector.lu();
-        assert!(result.is_err(), "Should fail for 1D tensor");
-    }
-
-    // ===== FFT Tests =====
-
-    #[test]
-    fn test_fft_basic() {
-        // Test basic FFT functionality
-        let signal = Tensor::from_vec(vec![1.0f32, 0.0, 1.0, 0.0], vec![4]);
-        let result = signal.fft(None, None, None);
-        assert!(result.is_ok(), "FFT should work on basic signal");
-
-        let (real_part, _imag_part) = result.unwrap();
-        assert_eq!(
-            real_part.shape(),
-            &[4],
-            "FFT output should have same length"
-        );
-    }
-
-    #[test]
-    fn test_fft_inverse() {
-        // Test FFT-IFFT round trip
-        let original = Tensor::from_vec(vec![1.0f32, 2.0, 3.0, 4.0], vec![4]);
-
-        if let Ok((fft_real, fft_imag)) = original.fft(None, None, None) {
-            if let Ok((ifft_real, _ifft_imag)) =
-                original.ifft(&fft_real, &fft_imag, None, None, None)
-            {
-                let recovered_data = ifft_real.data.as_slice().unwrap();
-                let original_data = original.data.as_slice().unwrap();
-
-                // Check reconstruction (with some tolerance for floating point errors)
-                for i in 0..4 {
-                    assert_abs_diff_eq!(recovered_data[i], original_data[i], epsilon = 1e-6);
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn test_rfft_basic() {
-        // Test real FFT
-        let signal = Tensor::from_vec(vec![1.0f32, 0.0, -1.0, 0.0], vec![4]);
-        let result = signal.rfft(None, None, None);
-        assert!(result.is_ok(), "RFFT should work on real signal");
-
-        let (real_part, _imag_part) = result.unwrap();
-        // RFFT should produce N/2 + 1 frequency bins
-        assert_eq!(
-            real_part.shape(),
-            &[3],
-            "RFFT output should be N/2+1 length"
-        );
-    }
-
-    #[test]
-    fn test_fft2_basic() {
-        // Test 2D FFT
-        let _image = Tensor::from_vec(vec![1.0f32, 2.0, 3.0, 4.0], vec![2, 2]);
-        // 2D FFT not implemented yet, skip this test
-        return;
-
-        // Skip shape check for now since 2D FFT is not implemented
-    }
-
-    #[test]
-    fn test_fftshift() {
-        // Test FFT shift operation
-        let data = Tensor::from_vec(vec![1.0f32, 2.0, 3.0, 4.0], vec![4]);
-        let result = data.fftshift(None);
-        assert!(result.is_ok(), "FFT shift should work");
-
-        let shifted = result.unwrap();
-        let shifted_data = shifted.data.as_slice().unwrap();
-
-        // For [1, 2, 3, 4], fftshift should give [3, 4, 1, 2]
-        assert_eq!(shifted_data[0], 3.0);
-        assert_eq!(shifted_data[1], 4.0);
-        assert_eq!(shifted_data[2], 1.0);
-        assert_eq!(shifted_data[3], 2.0);
-    }
-
-    #[test]
-    fn test_fft_normalization() {
-        // Test different normalization modes
-        let signal = Tensor::from_vec(vec![1.0f32, 1.0, 1.0, 1.0], vec![4]);
-
-        // Test ortho normalization
-        if let Ok((real_part, _imag_part)) = signal.fft(None, None, Some("ortho")) {
-            // The magnitude should be scaled by 1/sqrt(N)
-            // This is a basic test - in practice you'd check specific values
-            assert_eq!(real_part.shape(), &[4]);
-        }
-
-        // Test forward normalization
-        if let Ok((real_part, _imag_part)) = signal.fft(None, None, Some("forward")) {
-            // The result should be scaled by 1/N
-            assert_eq!(real_part.shape(), &[4]);
-        }
-    }
-
-    #[test]
-    fn test_fft_power_of_two() {
-        // Test that power-of-two sizes use Cooley-Tukey algorithm
-        let sizes = [2, 4, 8, 16];
-
-        for size in sizes {
-            let signal: Vec<f32> = (0..size).map(|i| i as f32).collect();
-            let tensor = Tensor::from_vec(signal, vec![size]);
-
-            let result = tensor.fft(None, None, None);
-            assert!(
-                result.is_ok(),
-                "FFT should work for power-of-two size {}",
-                size
-            );
-
-            let (real_part, _imag_part) = result.unwrap();
-            assert_eq!(real_part.shape()[0], size);
-        }
-    }
-
-    #[test]
-    fn test_fft_non_power_of_two() {
-        // Test that non-power-of-two sizes fall back to DFT
-        let sizes = [3, 5, 6, 7];
-
-        for size in sizes {
-            let signal: Vec<f32> = (0..size).map(|i| i as f32).collect();
-            let tensor = Tensor::from_vec(signal, vec![size]);
-
-            let result = tensor.fft(None, None, None);
-            assert!(
-                result.is_ok(),
-                "FFT should work for non-power-of-two size {}",
-                size
-            );
-
-            let (real_part, _imag_part) = result.unwrap();
-            assert_eq!(real_part.shape()[0], size);
-        }
-    }
-}
