--- src/tensor/zero_copy.rs
@@ -1,567 +1,0 @@
-//! # Zero-Copy Tensor Operations
-//! ゼロコピーテンソル操作
-//! 
-//! This module provides zero-copy tensor operations that eliminate unnecessary
-//! data copying through tensor views, shared ownership, and in-place operations,
-//! significantly improving memory efficiency and performance.
-//! 
-//! ## Overview
-//! 
-//! Zero-copy operations are essential for high-performance tensor computing:
-//! 
-//! - [`TensorView`]: Immutable zero-copy view of tensor data
-//! - [`TensorViewMut`]: Mutable zero-copy view for in-place operations
-//! - [`SharedTensor`]: Reference-counted shared tensor ownership
-//! - [`ZeroCopyOps`]: Trait for zero-copy tensor operations
-//! 
-//! ## Key Benefits
-//! 
-//! - **Memory Efficiency**: Eliminates data copying for view operations
-//! - **Performance**: Reduces memory allocation and deallocation overhead
-//! - **Cache Locality**: Better cache utilization through data reuse
-//! - **Memory Safety**: Rust's ownership system ensures safe shared access
-//! 
-//! ## Usage Examples
-//! 
-//! ### Basic Zero-Copy Views
-//! 
-//! ```rust
-//! use rustorch::tensor::Tensor;
-//! 
-//! let tensor = Tensor::<f32>::from_vec(vec![1.0, 2.0, 3.0, 4.0], vec![2, 2]);
-//! 
-//! // Basic tensor operations
-//! println!("Tensor shape: {:?}", tensor.shape());
-//! 
-//! // Create mutable tensor
-//! let mut tensor_mut = Tensor::<f32>::zeros(&[2, 2]);
-//! println!("Mutable tensor created");
-//! ```
-//! 
-//! ### Shared Tensor Operations
-//! 
-//! ```rust
-//! use rustorch::tensor::Tensor;
-//! 
-//! let tensor = Tensor::<f32>::ones(&[100, 100]);
-//! 
-//! // Multiple tensors can share data efficiently
-//! let tensor1 = tensor.clone(); // Reference-counted sharing
-//! let tensor2 = tensor.clone();
-//! 
-//! // All tensors reference the same underlying data
-//! assert_eq!(tensor1.shape(), tensor2.shape());
-//! assert_eq!(tensor.shape(), tensor1.shape());
-//! ```
-//! 
-//! ### Zero-Copy Element-wise Operations
-//! 
-//! ```rust
-//! use rustorch::tensor::Tensor;
-//! 
-//! let tensor1 = Tensor::<f32>::ones(&[1000, 1000]);
-//! let tensor2 = Tensor::<f32>::ones(&[1000, 1000]);
-//! 
-//! // Element-wise operations 
-//! let result = &tensor1 + &tensor2;
-//! 
-//! # assert_eq!(result.shape(), &[1000, 1000]);
-//! # Ok::<(), Box<dyn std::error::Error>>(())
-//! ```
-//! 
-//! ### In-Place Operations
-//! 
-//! ```rust
-//! use rustorch::tensor::Tensor;
-//! 
-//! let mut tensor = Tensor::<f32>::zeros(&[1000, 1000]);
-//! let other = Tensor::<f32>::ones(&[1000, 1000]);
-//! 
-//! // In-place operations modify tensor without allocation
-//! tensor.inplace_add(&other)?;
-//! tensor.inplace_mul_scalar(2.0)?;
-//! tensor.inplace_apply(|x| x.sin())?;
-//! 
-//! // Verify in-place modification
-//! assert!(tensor.iter().all(|&x| x > 0.0));
-//! # Ok::<(), Box<dyn std::error::Error>>(())
-//! ```
-//! 
-//! ### Memory-Efficient Tensor Slicing
-//! 
-//! ```rust
-//! use rustorch::tensor::Tensor;
-//! 
-//! let large_tensor = Tensor::<f32>::ones(&[1000, 1000]);
-//! 
-//! // Create zero-copy slice views
-//! let top_half = large_tensor.slice_view(&[0..500, 0..1000])?;
-//! let bottom_half = large_tensor.slice_view(&[500..1000, 0..1000])?;
-//! 
-//! // Process slices independently without copying data
-//! let top_sum = top_half.sum();
-//! let bottom_sum = bottom_half.sum();
-//! 
-//! println!("Top half sum: {}, Bottom half sum: {}", top_sum, bottom_sum);
-//! # Ok::<(), Box<dyn std::error::Error>>(())
-//! ```
-//! 
-//! ## Performance Characteristics
-//! 
-//! - **Zero Memory Overhead**: Views add no memory allocation overhead
-//! - **Cache Efficiency**: Better cache utilization through data locality
-//! - **Reduced GC Pressure**: Fewer allocations reduce garbage collection overhead
-//! - **Vectorization Friendly**: Contiguous memory layout enables SIMD optimization
-//! 
-//! ## Memory Safety Guarantees
-//! 
-//! - **Borrow Checker**: Rust's borrow checker prevents data races
-//! - **Lifetime Management**: Views cannot outlive their source tensors
-//! - **Thread Safety**: SharedTensor provides thread-safe shared ownership
-//! - **Mutation Safety**: Mutable views ensure exclusive access to data
-//! 
-//! ## Integration with Other Modules
-//! 
-//! Zero-copy operations integrate seamlessly with:
-//! 
-//! - **Parallel Operations**: Zero-copy views work with parallel traits
-//! - **SIMD Operations**: Aligned views enable vectorized operations
-//! - **GPU Operations**: Views can be transferred to GPU without copying
-//! - **Memory Pools**: Views work with pooled memory allocation
-
-use super::Tensor;
-use crate::error::{RusTorchError, RusTorchResult};
-use ndarray::{ArrayD, ArrayViewD, ArrayViewMutD};
-use num_traits::Float;
-use std::sync::Arc;
-
-/// Zero-copy tensor view for read-only operations
-/// 読み取り専用操作用のゼロコピーテンソルビュー
-pub struct TensorView<'a, T: Float> {
-    view: ArrayViewD<'a, T>,
-}
-
-/// Zero-copy mutable tensor view for in-place operations
-/// インプレース操作用のゼロコピー可変テンソルビュー
-pub struct TensorViewMut<'a, T: Float> {
-    view: ArrayViewMutD<'a, T>,
-}
-
-impl<'a, T: Float + 'static> TensorView<'a, T> {
-    /// Create a new tensor view from array view
-    /// 配列ビューから新しいテンソルビューを作成
-    pub fn new(view: ArrayViewD<'a, T>) -> Self {
-        Self { view }
-    }
-
-    /// Get the shape of the tensor view
-    /// テンソルビューの形状を取得
-    pub fn shape(&self) -> &[usize] {
-        self.view.shape()
-    }
-
-    /// Get element at index
-    /// インデックスの要素を取得
-    pub fn get(&self, index: &[usize]) -> Option<T> {
-        self.view.get(index).copied()
-    }
-
-    /// Convert to owned tensor (requires allocation)
-    /// 所有テンソルに変換（割り当てが必要）
-    pub fn to_owned(&self) -> Tensor<T> {
-        Tensor {
-            data: self.view.to_owned(),
-        }
-    }
-
-    /// Element-wise operation with another view (zero-copy)
-    /// 他のビューとの要素ごと演算（ゼロコピー）
-    pub fn elementwise_with<F>(&self, other: &TensorView<'a, T>, op: F) -> ParallelResult<Tensor<T>>
-    where
-        F: Fn(T, T) -> T + Send + Sync,
-        T: Send + Sync,
-    {
-        if self.view.shape() != other.view.shape() {
-            return Err(RusTorchError::parallel("Shape mismatch"));        }
-
-        let mut result = Tensor::zeros(self.view.shape());
-        
-        if let (Some(self_slice), Some(other_slice), Some(result_slice)) = (
-            self.view.as_slice(),
-            other.view.as_slice(),
-            result.data.as_slice_mut()
-        ) {
-            result_slice.iter_mut()
-                .zip(self_slice.iter())
-                .zip(other_slice.iter())
-                .for_each(|((r, &a), &b)| {
-                    *r = op(a, b);
-                });
-        }
-
-        Ok(result)
-    }
-
-    /// Reduction operation (zero-copy input)
-    /// リダクション演算（ゼロコピー入力）
-    pub fn reduce<F>(&self, init: T, op: F) -> T
-    where
-        F: Fn(T, T) -> T,
-    {
-        self.view.iter().fold(init, |acc, &x| op(acc, x))
-    }
-}
-
-impl<'a, T: Float> TensorViewMut<'a, T> {
-    /// Create a new mutable tensor view
-    /// 新しい可変テンソルビューを作成
-    pub fn new(view: ArrayViewMutD<'a, T>) -> Self {
-        Self { view }
-    }
-
-    /// Get the shape of the tensor view
-    /// テンソルビューの形状を取得
-    pub fn shape(&self) -> &[usize] {
-        self.view.shape()
-    }
-
-    /// In-place element-wise operation (true zero-copy)
-    /// インプレース要素ごと演算（真のゼロコピー）
-    pub fn elementwise_inplace<F>(&mut self, other: &TensorView<T>, op: F) -> ParallelResult<()>
-    where
-        F: Fn(T, T) -> T + Send + Sync,
-        T: Send + Sync,
-    {
-        if self.view.shape() != other.view.shape() {
-            return Err(RusTorchError::parallel("Shape mismatch"));        }
-
-        if let (Some(self_slice), Some(other_slice)) = (
-            self.view.as_slice_mut(),
-            other.view.as_slice()
-        ) {
-            self_slice.iter_mut()
-                .zip(other_slice.iter())
-                .for_each(|(a, &b)| {
-                    *a = op(*a, b);
-                });
-        }
-
-        Ok(())
-    }
-
-    /// In-place scalar operation (true zero-copy)
-    /// インプレーススカラー演算（真のゼロコピー）
-    pub fn scalar_inplace<F>(&mut self, scalar: T, op: F)
-    where
-        F: Fn(T, T) -> T,
-    {
-        if let Some(slice) = self.view.as_slice_mut() {
-            slice.iter_mut().for_each(|x| {
-                *x = op(*x, scalar);
-            });
-        }
-    }
-
-    /// Fill with value (zero-copy)
-    /// 値で埋める（ゼロコピー）
-    pub fn fill(&mut self, value: T) {
-        self.view.fill(value);
-    }
-}
-
-/// Zero-copy operations for Tensor
-/// Tensor用ゼロコピー演算
-impl<T: Float + Clone + Send + Sync + 'static> Tensor<T> {
-    /// Create a zero-copy read-only view of the tensor
-    /// テンソルのゼロコピー読み取り専用ビューを作成
-    pub fn zero_copy_view(&self) -> TensorView<T> {
-        TensorView::new(self.data.view())
-    }
-
-    /// Create a zero-copy mutable view of the tensor
-    /// テンソルのゼロコピー可変ビューを作成
-    pub fn zero_copy_view_mut(&mut self) -> TensorViewMut<T> {
-        TensorViewMut::new(self.data.view_mut())
-    }
-
-    /// Slice tensor without copying data
-    /// データをコピーせずにテンソルをスライス
-    pub fn slice_view(&self, ranges: &[std::ops::Range<usize>]) -> ParallelResult<Tensor<T>> {
-        if ranges.len() != self.data.ndim() {
-            return Err(RusTorchError::parallel("Dimension error"));        }
-
-        // Validate ranges
-        for (i, range) in ranges.iter().enumerate() {
-            if range.end > self.data.shape()[i] {
-            return Err(RusTorchError::parallel("Dimension error"));            }
-        }
-
-        // Create actual sliced tensor by extracting data manually
-        let mut result_shape = Vec::new();
-        let mut result_data = Vec::new();
-        
-        // Calculate result shape
-        for (_i, range) in ranges.iter().enumerate() {
-            result_shape.push(range.end - range.start);
-        }
-        
-        // Extract sliced data manually
-        let result_size: usize = result_shape.iter().product();
-        result_data.reserve(result_size);
-        
-        // For 2D case (most common), handle directly
-        if ranges.len() == 2 {
-            let r0 = &ranges[0];
-            let r1 = &ranges[1];
-            let original_shape = self.data.shape();
-            
-            for i in r0.start..r0.end {
-                for j in r1.start..r1.end {
-                    let idx = i * original_shape[1] + j;
-                    if let Some(val) = self.data.as_slice().unwrap().get(idx) {
-                        result_data.push(*val);
-                    }
-                }
-            }
-        } else {
-            // For other dimensions, use a more general approach
-            use ndarray::Slice;
-            let mut current_array = self.data.clone();
-            for (axis, range) in ranges.iter().enumerate() {
-                let sliced = current_array.slice_axis(ndarray::Axis(axis), Slice::from(range.start..range.end));
-                current_array = sliced.to_owned();
-            }
-            return Ok(Tensor::new(current_array));
-        }
-        
-        // Create new tensor from extracted data
-        let sliced_array = ndarray::Array::from_shape_vec(result_shape, result_data)
-            .map_err(|_| RusTorchError::parallel("Shape mismatch")?;
-        Ok(Tensor::new(sliced_array.into_dyn()))
-    }
-
-    /// Reshape tensor view without copying data
-    /// データをコピーせずにテンソルビューをリシェイプ
-    pub fn reshape_view(&self, new_shape: &[usize]) -> ParallelResult<Tensor<T>> {
-        let total_elements: usize = new_shape.iter().product();
-        if total_elements != self.data.len() {
-            return Err(RusTorchError::parallel("Shape mismatch"));        }
-
-        // Create reshaped tensor using ndarray reshape
-        let reshaped_array = self.data.view()
-            .into_shape(new_shape)
-            return Err(RusTorchError::parallel("Shape mismatch"));        }
-
-        // Check shape compatibility
-        for view in &views[1..] {
-            let shape = view.shape();
-            if shape.len() != first_shape.len() {
-            return Err(RusTorchError::parallel("Shape mismatch"));            }
-            for (i, (&dim1, &dim2)) in first_shape.iter().zip(shape.iter()).enumerate() {
-                if i != axis && dim1 != dim2 {
-            return Err(RusTorchError::parallel("Shape mismatch"));                }
-            }
-        }
-
-        // Calculate result shape
-        let mut result_shape = first_shape.to_vec();
-        result_shape[axis] = views.iter().map(|v| v.shape()[axis]).sum();
-
-        // Create result tensor and copy data
-        let mut result = Tensor::zeros(&result_shape);
-        let mut offset = 0;
-
-        for view in views {
-            let view_size = view.shape()[axis];
-            // Copy data from each view
-            // This is a simplified implementation - real implementation would need proper indexing
-            if let (Some(view_slice), Some(result_slice)) = (
-                view.view.as_slice(),
-                result.data.as_slice_mut()
-            ) {
-                let start = offset * view_slice.len() / view_size;
-                let end = start + view_slice.len();
-                if end <= result_slice.len() {
-                    result_slice[start..end].copy_from_slice(view_slice);
-                }
-            }
-            offset += view_size;
-        }
-
-        Ok(result)
-    }
-
-    /// Memory-mapped tensor operations (placeholder for future implementation)
-    /// メモリマップテンソル演算（将来の実装用プレースホルダー）
-    pub fn from_memory_map(_file_path: &str) -> ParallelResult<Self> {
-        // Placeholder for memory-mapped tensor creation
-        // 実際の実装では、ファイルをメモリマップしてテンソルを作成
-        Err(RusTorchError::parallel(parallel_execution_error(
-            "Memory mapping not yet implemented"
-        ).into())
-    }
-}
-
-/// Shared tensor for zero-copy sharing between threads
-/// スレッド間でのゼロコピー共有用共有テンソル
-pub struct SharedTensor<T: Float> {
-    data: Arc<ArrayD<T>>,
-}
-
-impl<T: Float + Clone + Send + Sync + 'static> SharedTensor<T> {
-    /// Create shared tensor from regular tensor
-    /// 通常のテンソルから共有テンソルを作成
-    pub fn new(tensor: Tensor<T>) -> Self {
-        Self {
-            data: Arc::new(tensor.data),
-        }
-    }
-
-    /// Get read-only view of shared tensor
-    /// 共有テンソルの読み取り専用ビューを取得
-    pub fn view(&self) -> TensorView<T> {
-        TensorView::new(self.data.view())
-    }
-
-    /// Clone the shared reference (zero-copy)
-    /// 共有参照をクローン（ゼロコピー）
-    pub fn clone_ref(&self) -> Self {
-        Self {
-            data: Arc::clone(&self.data),
-        }
-    }
-
-    /// Convert back to owned tensor (requires cloning data)
-    /// 所有テンソルに戻す（データのクローンが必要）
-    pub fn to_owned(&self) -> Tensor<T> {
-        Tensor {
-            data: (*self.data).clone(),
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_tensor_view_creation() {
-        let tensor = Tensor::<f32>::ones(&[3, 4]);
-        let view = tensor.view();
-        
-        assert_eq!(view.shape(), &[3, 4]);
-        assert_eq!(view.get([0, 0]), Some(&1.0));
-    }
-
-    #[test]
-    fn test_mutable_view_operations() {
-        let mut tensor = Tensor::<f32>::zeros(&[2, 3]);
-        {
-            let view = tensor.view_mut();
-            view.fill(5.0);
-        }
-        
-        // Check that the original tensor was modified
-        assert_eq!(tensor.as_array()[[0, 0]], 5.0);
-        assert_eq!(tensor.as_array()[[1, 2]], 5.0);
-    }
-
-    #[test]
-    fn test_zero_copy_elementwise() {
-        let tensor1 = Tensor::<f32>::ones(&[2, 2]);
-        let tensor2 = Tensor::<f32>::ones(&[2, 2]);
-        
-        let view1 = tensor1.zero_copy_view();
-        let view2 = tensor2.zero_copy_view();
-        
-        let result = view1.elementwise_with(&view2, |a, b| a + b);
-        assert!(result.is_ok());
-        
-        let result = result.unwrap();
-        assert_eq!(result.as_array()[[0, 0]], 2.0);
-    }
-
-    #[test]
-    fn test_inplace_operations() {
-        let mut tensor1 = Tensor::<f32>::ones(&[2, 2]);
-        let tensor2 = Tensor::<f32>::ones(&[2, 2]);
-        
-        {
-            let mut view1 = tensor1.zero_copy_view_mut();
-            let view2 = tensor2.zero_copy_view();
-            
-            let result = view1.elementwise_inplace(&view2, |a, b| a + b);
-            assert!(result.is_ok());
-        }
-        
-        // Check that tensor1 was modified in place
-        assert_eq!(tensor1.as_array()[[0, 0]], 2.0);
-        assert_eq!(tensor1.as_array()[[1, 1]], 2.0);
-    }
-
-    #[test]
-    fn test_slice_view() {
-        let tensor = Tensor::<f32>::from_vec(
-            vec![1.0, 2.0, 3.0, 4.0, 5.0, 6.0],
-            vec![2, 3]
-        );
-        
-        let ranges = vec![0..1, 1..3];
-        let slice_view = tensor.slice_view(&ranges);
-        
-        assert!(slice_view.is_ok());
-        let slice_view = slice_view.unwrap();
-        assert_eq!(slice_view.shape(), &[1, 2]);
-    }
-
-    #[test]
-    fn test_reshape_view() {
-        let tensor = Tensor::<f32>::ones(&[2, 3]);
-        let reshaped = tensor.reshape_view(&[3, 2]);
-        
-        assert!(reshaped.is_ok());
-        let reshaped = reshaped.unwrap();
-        assert_eq!(reshaped.shape(), &[3, 2]);
-    }
-
-    #[test]
-    fn test_shared_tensor() {
-        let tensor = Tensor::<f32>::ones(&[2, 2]);
-        let shared = SharedTensor::new(tensor);
-        
-        let shared_clone = shared.clone_ref();
-        let view1 = shared.view();
-        let view2 = shared_clone.view();
-        
-        assert_eq!(view1.get(&[0, 0]), view2.get(&[0, 0]));
-    }
-
-    #[test]
-    fn test_reduction_operations() {
-        let tensor = Tensor::<f32>::from_vec(vec![1.0, 2.0, 3.0, 4.0], vec![2, 2]);
-        let view = tensor.view();
-        
-        let sum = view.iter().fold(0.0, |acc, &x| acc + x);
-        assert_eq!(sum, 10.0);
-        
-        let max = view.iter().fold(f32::NEG_INFINITY, |acc, &x| acc.max(x));
-        assert_eq!(max, 4.0);
-    }
-
-    #[test]
-    fn test_shape_mismatch_errors() {
-        let tensor1 = Tensor::<f32>::ones(&[2, 3]);
-        let tensor2 = Tensor::<f32>::ones(&[3, 2]);
-        
-        let view1 = tensor1.zero_copy_view();
-        let view2 = tensor2.zero_copy_view();
-        
-        let result = view1.elementwise_with(&view2, |a, b| a + b);
-        assert!(result.is_err());
-        
-        match result.unwrap_err() {
-            RusTorchError::parallel(ShapeMismatch { .. } => {},
-            _ => panic!("Expected ShapeMismatch error"),
-        }
-    }
-}
